% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_util.R
\name{ac}
\alias{ac}
\title{Adjust the transparency of a color or color palette.}
\usage{
ac(col, alpha = 0.5)
}
\arguments{
\item{col}{A (required) color or color palette (as a vector).}

\item{alpha}{A factor modifying the opacity alpha 
(as \code{alpha.f} in \code{\link{adjustcolor}}) to a value in \code{[0, 1]}. 
Default: \code{alpha = .50} (i.e., medium opacity).}
}
\value{
A color vector of the same length as \code{col}, 
transformed by \code{link{adjustcolor}}.
}
\description{
\code{ac} adjusts the transparency of a color or color palette \code{col} 
to an opacity level \code{alpha}.
}
\details{
\code{ac} is merely a convenient wrapper for 
\code{\link{adjustcolor}} of the \strong{grDevices} package.
}
\examples{
ac("black")  # using alpha = .5 by default

blacks <- c(ac("black", .25), ac("black"), ac("black", .75))
seecol(blacks)

seecol(ac(pal_unikn_pref, .67), title = "Adding color transparency by ac()")
 
}
\seealso{
\code{\link{seecol}} to plot color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{defpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other color functions: 
\code{\link{grepal}()},
\code{\link{newpal}()},
\code{\link{seecol}()},
\code{\link{shades_of}()},
\code{\link{usecol}()}
}
\concept{color functions}
