% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{Bordeaux}
\alias{Bordeaux}
\title{uni.kn color bordeaux (as df).}
\format{An object of class \code{character} of length 1.}
\usage{
Bordeaux
}
\description{
\code{bordeaux} provides the preferred color of \code{\link{pal_bordeaux}} 
(as an atomic data frame) and is defined as 
\code{\link{pal_bordeaux}[4]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
bordeaux  # name "bordeaux4" and color value "#8E2043" (as df)
all.equal(bordeaux, pal_bordeaux[4])  # TRUE (both df)

Bordeaux  # HEX character "#8E2043" (as value)
all.equal(Bordeaux, pal_bordeaux[[4]])  # TRUE (same HEX values)

}
\seealso{
\code{\link{pal_bordeaux}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{seepal}} to show and use color palettes. 
uni.kn color Bordeaux.    

\code{Bordeaux} provides the preferred color of \code{\link{pal_bordeaux}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_bordeaux}[[4]]}. 

See \url{https://www.uni-konstanz.de} for details.

\code{\link{pal_bordeaux}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{seepal}} to show and use color palettes.

Other named colors (as df): \code{\link{Grau}},
  \code{\link{Karpfenblau}}, \code{\link{Peach}},
  \code{\link{Petrol}}, \code{\link{Pinky}},
  \code{\link{Seeblau}}, \code{\link{Seegruen}},
  \code{\link{Signal}}

Other preferred colors: \code{\link{Grau}},
  \code{\link{Karpfenblau}}, \code{\link{Peach}},
  \code{\link{Petrol}}, \code{\link{Pinky}},
  \code{\link{Seeblau}}, \code{\link{Seegruen}},
  \code{\link{Signal}}
}
\concept{named colors (as df)}
\concept{preferred colors}
\keyword{datasets}
