% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text.R
\name{box_text}
\alias{box_text}
\title{\code{box_text} places 1 or more text strings (of a character vector \code{lbls}) 
onto a plot and places a colored box behind
each label to highlight it (i.e., make it stand out from the background).}
\usage{
box_text(x, y, lbls = NA, col_lbl = NULL, col_bg = NA,
  col_bg_brd = NA, adj = NULL, pos = NULL, offset = 0.5,
  padding = c(0.5, 0.5), cex = 1, font = 2)
}
\arguments{
\item{x}{A numeric vector of x-coordinates at which the 
text labels in \code{lbls} should be written. 
If the length of \code{x} and \code{y} differs, the shorter one is recycled.}

\item{y}{A numeric vector of y-coordinates at which the 
text labels in \code{lbls} should be written.}

\item{lbls}{A character vector specifying the text labels 
to be written.}

\item{col_lbl}{The color(s) of the text label(s).}

\item{col_bg}{The color(s) to fill or shade the rectangle(s) with. 
The default \code{col_bg = NA} implies no fill 
(i.e., draws transparent rectangles).}

\item{col_bg_brd}{The color(s) for rectangle border(s). 
The default \code{NA} omits borders.}

\item{adj}{One or two adjustment values in \code{[0, 1]} which specify 
the x (and optionally y) adjustment of the labels.}

\item{pos}{A position specifier for the text. If specified this overrides
any \code{adj} value given. Values of 1, 2, 3 and 4, respectively indicate
positions below, to the left of, above and to the right of the specified
coordinates.}

\item{offset}{When \code{pos} is specified, this value sets the offset of
the label from the specified coordinate in fractions of a character width.}

\item{padding}{A factor used for the padding of the box around
the text. Padding is specified in fractions of a character width. If a
vector of length 2 is specified then 2 different factors are used for the
padding in x- and y-direction.}

\item{cex}{A numeric character expansion factor, which is
multiplied by \code{par("cex")} to yield the final character size.}

\item{font}{The font to be used. 
Default: \code{font = 2} (bold).}
}
\value{
Returns the coordinates of the background rectangle(s). 
If multiple labels are placed in a vector then the coordinates are returned
as a matrix with columns corresponding to xleft, xright, ybottom, ytop.
If only one label is placed, its coordinates are returned as a vector.
}
\description{
\code{box_text} uses the base graphics system \code{graphics::}.
}
\examples{
## Create some noisy background:
plot(x = runif(500), y = runif(500), type = "p", pch = 16, cex = 3, col = grey(.33, .10))

## Adjust cex, font and adj:
box_text(x = .01, y = .95, lbls = "What a messy plot",
         col_bg = "skyblue1", adj = c(0, 0), padding = c(.25, .75), cex = 1.8)

## Vector of 2 lbls, using argument 'pos' to position right of coordinates:
box_text(x = c(.15, .60), y = c(.10, .50),
         lbls = c("Some highlighted text here.", "Something else here."),
         col_bg = c("skyblue2", "gold"), pos = 4, padding = c(.25, 1))
   
}
\author{
Adapted from code by Ian Kopacka.
}
