% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{col_n}
\alias{col_n}
\title{Get n dedicated colors of a color palette.}
\usage{
col_n(n = "all", pal = pal_unikn)
}
\arguments{
\item{n}{A number specifying the desired number colors of pal (as a number) 
or the character string \code{"all"} (to get all colors of \code{pal}). 
Default: \code{n = "all"}.}

\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{col_n} returns \code{n} dedicated colors of a color palette \code{pal}.
}
\examples{
col_n()    # all default colors

# Default palette pal:
col_n(1)   # n = 1 color of default color pal
col_n(2)
col_n(4)
col_n(6)
col_n(8)
col_n(10)  # n = 10 colors of default color pal

col_n(11)  # return maximal number of colors

# Other palettes pal:
col_n(3,  RColorBrewer::brewer.pal(n = 5, name = "Blues"))
col_n(10, RColorBrewer::brewer.pal(n = 5, name = "Blues"))

}
\seealso{
\code{\link{col_n_sq}} for \code{n^2} dedicated colors of a color palette; 
\code{\link{col_which}} for specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color scales: \code{\link{col_n_sq}},
  \code{\link{col_which}}, \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn}}, \code{\link{plot_pal}}
}
\concept{color scales}
