% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{col_n_sq}
\alias{col_n_sq}
\title{Get n^2 dedicated colors of a color palette.}
\usage{
col_n_sq(n = "all", pal = pal_unikn)
}
\arguments{
\item{n}{A number specifying the desired number colors of pal (as a number) 
or the character string \code{"all"} (to get all colors of \code{pal}). 
Default: \code{n = "all"}.}

\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{col_n_sq} returns \code{n^2} dedicated colors of a color palette \code{pal} 
(up to a maximum of \code{n + 1} colors).
}
\examples{
col_n_sq(1)
col_n_sq(2)
col_n_sq(3) #  9 colors
col_n_sq(4) # 11 colors = maximum of (n + 1) colors.

}
\seealso{
\code{\link{col_n}} for \code{n} dedicated colors of a color palette; 
\code{\link{col_which}} for specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color scales: \code{\link{col_n}},
  \code{\link{col_which}}, \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn}}, \code{\link{plot_pal}}
}
\concept{color scales}
