% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{col_scale}
\alias{col_scale}
\title{Provide a color gradient scale.}
\usage{
col_scale(col = pal_unikn, n = length(col), ...)
}
\arguments{
\item{col}{A list of colors or color palettes. 
Default: \code{col = \link{pal_unikn}}.}

\item{...}{Additional parameters 
(passed to \code{grDevices::colorRampPalette}).}
}
\description{
\code{col_scale} provides a color gradient 
based on a sequence of input colors or color palettes 
defined by \code{col}.
}
\details{
\code{col_scale} uses \code{grDevices::colorRampPalette} to  
provide a function that takes an integer argument 
(the required number of colors) to  
return a character vector of (rgb) colors 
interpolating color sequence provided by \code{col}.

For best results, consider combining existing color palettes 
with base color "white" (see Examples).
}
\examples{
## Creating color gradients: ------ 

# Gradients extending 1 color palette:
col_scale()(5)  # 5 colors of default scale pal_unikn
col_scale(pal_petrol)(10)  # 10 colors of pal_petrol

# Gradients over 2+ colors:
col_scale(c(Seeblau, Peach))(10)
col_scale(c(Signal, Petrol))(10)  
col_scale(c(Seeblau, "white", Pinky))(10)
col_scale(c(Karpfenblau, Seeblau, "gold"))(10)  # "gold" shines brighter than signal

# Gradients over 2+ color palettes:
col_scale(c(pal_seeblau, pal_peach))(10)
col_scale(c(pal_seeblau, pal_karpfenblau))(10)
col_scale(c(pal_seeblau, pal_grau, pal_petrol))(10)

# Gradients over 3+ color palettes and colors:
col_scale(c(rev(pal_seeblau), "white", pal_pinky))(11)
col_scale(c(rev(pal_seeblau), "white", pal_petrol))(11)
col_scale(c(rev(pal_karpfenblau), "white", pal_bordeaux))(11) 

## Creating and plotting color gradients: ------ 

# (1) Extending color palettes: ----
seecol(col_scale()(20))
seecol(col_scale(pal_seeblau)(10))
seecol(col_scale(pal_bordeaux)(10))

# (2) Combining colors or palettes: ---- 

# Combining 2+ colors to create new palettes:
seecol(col_scale(c(Seeblau, Signal))(10))
seecol(col_scale(c(Signal, Petrol))(10))
seecol(col_scale(c(Seeblau, "white", Pinky))(10))
seecol(col_scale(c(Karpfenblau, Seeblau, "white"))(10))
seecol(col_scale(c(Bordeaux, "white", Petrol))(10))
seecol(col_scale(c(Seeblau, "white", Petrol))(10))
seecol(col_scale(c(Karpfenblau, Seeblau, "gold"))(10))  # "gold" shines brighter than signal
  
# Combining 2+ color palettes into new palettes:
seecol(col_scale(c(rev(pal_seeblau), pal_petrol))(10))
seecol(col_scale(c(rev(pal_seeblau), pal_peach))(10))
seecol(col_scale(c(rev(pal_karpfenblau), pal_bordeaux))(10))
seecol(col_scale(c(rev(pal_seegruen), pal_pinky))(10))
seecol(col_scale(c(rev(pal_petrol), pal_bordeaux))(10))

#' # Combining 2 palettes and mid-color "white" to new palettes:
seecol(col_scale(c(rev(pal_seeblau), "white", pal_pinky))(11))
seecol(col_scale(c(rev(pal_seeblau), "white", pal_petrol))(11))
seecol(col_scale(c(rev(pal_karpfenblau), "white", pal_bordeaux))(11))
seecol(col_scale(c(rev(pal_seegruen), "white", pal_peach))(11))
seecol(col_scale(c(rev(pal_petrol), "white", pal_bordeaux))(11))

}
\seealso{
\code{\link{seecol}} to plot color palettes; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{pal_which}} to get specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{pal_bordeaux}},
  \code{\link{pal_grau}}, \code{\link{pal_karpfenblau}},
  \code{\link{pal_n_sq}}, \code{\link{pal_n}},
  \code{\link{pal_peach}}, \code{\link{pal_petrol}},
  \code{\link{pal_pinky}}, \code{\link{pal_seeblau}},
  \code{\link{pal_seegruen}}, \code{\link{pal_signal}},
  \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{plot_pal}}, \code{\link{seecol}}
}
\concept{color palettes}
