% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{col_which}
\alias{col_which}
\title{Get specific colors of a color palette.}
\usage{
col_which(which = "all", pal = pal_unikn)
}
\arguments{
\item{which}{A numeric vector specifying the desired colors of pal (by position) 
or the character string \code{"all"} (to get all colors of \code{pal}). 
Default: \code{which = "all"}.}

\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{col_which} returns the colors of a color palette \code{pal} 
that are specified in a numeric vector \code{which}.
}
\examples{
col_which()  # returns default color palette
col_which(which = "all")
col_which(which = 1:3)
col_which(which = rep(1, 20))

}
\seealso{
\code{\link{col_n}} for \code{n} dedicated colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color scales: \code{\link{col_n_sq}},
  \code{\link{col_n}}, \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn}}, \code{\link{plot_pal}}
}
\concept{color scales}
