% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demo.R
\name{demopal}
\alias{demopal}
\alias{democol}
\alias{demofun}
\title{Demonstrate a color palette.}
\usage{
demopal(pal = pal_unikn, type = NA, ...)
}
\arguments{
\item{pal}{A color palette (to be illustrated).
Default: \code{pal = pal_unikn}.}

\item{type}{The type of plot to be used (as character or integer index). 
Permissible types are \code{"bar"}, \code{"mosaic"}, \code{"scatter"} (or 1 to 3).}
}
\value{
The random data that was plotted (as an invisible R object).
}
\description{
\code{demopal} provides an example plot of some \code{type} 
to illustrate a color palette \code{pal}.
}
\details{
The \code{demopal} wrapper function passes a range of arguments to more specific functions. 
Common arguments include:

\itemize{

  \item \code{col_par} Default color for \code{par(col)}; 
  \item \code{alpha} Default value for color transparency (in 0:1);  
  \item \code{n} A scaling parameter (for random data generation);  

  \item \code{main} plot title (on top); 
  \item \code{sub} subtitle (on right margin);  
  
  \item \code{pal_name} The name of input color palette \code{pal}; 
  \item \code{seed} A random seed (for reproducible randomness). 
  
}

Some functions additionally accept type-specific arguments 
(e.g., \code{beside}, \code{horiz}, and \code{as_prop}, for plot \code{type = "bar"}, 
and \code{cex} for plot \code{type = "scatter"}). 

The type-specific functions usually generate some random data 
(scaled by a parameter \code{n}) that is being plotted. 
This data is returned (as an invisible R object)  
to enable a plot's reconstruction.
}
\examples{
demopal(pal_petrol, type = 1)
demopal(c(rev(pal_pinky), pal_seeblau), type = 2)
demopal(type = "scatter", col_par = "black", n = 100, cex = 10, seed = 101)

}
\seealso{
\code{\link{seepal}} for plotting color palettes;  
\code{\link{usecol}} for using color palettes; 
\code{\link{shades_of}} to defining shades of a given color; 
\code{\link{ac}} for adjusting color transparency; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color functions: 
\code{\link{ac}()},
\code{\link{grepal}()},
\code{\link{newpal}()},
\code{\link{seecol}()},
\code{\link{shades_of}()},
\code{\link{simcol}()},
\code{\link{usecol}()}
}
\concept{color functions}
