% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{eth_pal_light}
\alias{eth_pal_light}
\title{Lighter shades of the ETH Zurich colors}
\format{
An object of class \code{character} of length 7.
}
\source{
Color definitions are based on 
\url{https://ethz.ch/staffnet/en/service/communication/corporate-design/digital-media/web-colours.html}.
}
\usage{
eth_pal_light
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{eth_pal_light} provides the lighter shades of the color palette 
of the \href{https://ethz.ch/}{ETH Zürich}, Switzerland.
}
\examples{
eth_pal_light
seecol(eth_pal_light)  # view color palette
demopal(eth_pal_light, type = 4, main = "Light shades of the ETH Zürich colors")

}
\seealso{
\code{\link{eth_pal}} and \code{\link{eth_pal_grey}} for alternative colors of the ETH Zurich;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_grey}},
\code{\link{eth_pal}},
\code{\link{mpg_pal}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}}
}
\author{
Hansjoerg Neth, 2022-10-25.
}
\concept{contributed color palettes}
\keyword{datasets}
