% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{frame}
\alias{frame}
\title{Plot a frame (or slide).}
\usage{
frame(col = NA, dim = c(4/3, 1), border = grey(0.33, 1), lwd = 1.5)
}
\arguments{
\item{col}{The color to fill the frame with (i.e., its background color).
Default: \code{col = NA} (i.e., system default for transparency).}

\item{dim}{The x- and y-dimensions of the frame. 
Default: \code{dim = c(4/3, 1)} (i.e., unit height, 4/3 wider than high).}

\item{border}{The color of the border. 
Default: \code{border = grey(.33, 1)}.}

\item{lwd}{The line width of the border. 
Default: \code{lwd = 1.5}.}
}
\description{
\code{frame} plots an empty frame or slide.
}
\examples{
frame()  # default frame

}
\seealso{
\code{\link{xbox}} to plot a box.

Other plot functions: \code{\link{xbox}}
}
\concept{plot functions}
