% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{fu_pal_0}
\alias{fu_pal_0}
\title{Primary colors of the Free University Berlin}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on 
\url{https://www.fu-berlin.de/sites/corporate-design/grundlagen/farben/index.html}.
}
\usage{
fu_pal_0
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{fu_pal_0} provides the two primary colors 
of the \href{https://www.fu-berlin.de/en/index.html}{Free University Berlin}, Germany.
}
\details{
The dark \code{"FU blue"} implies consistency, seriosity and respect. 
It is defined as Pantone 280C 
and approximated by CMYK 100/72/0/18.5, RGB 0/51/102, and HEX #003366.

The bright \code{"FU green"} implies renewal and mobility. 
It is is defined as Pantone 381C 
and approximated by CMYK 18.5/0/91/0, RGB 153/204/0, and HEX #99CC00. 

See \code{\link{fu_pal_1}} for a corresponding color gradient.
}
\examples{
fu_pal_0
seecol(fu_pal_0, main = "Primary colors of the FU Berlin")
demopal(fu_pal_0, type = 1, main = "Primary colors of the FU Berlin")

}
\seealso{
\code{\link{fu_pal_1}} for a corresponding color gradient of the FU Berlin;
\code{\link{fu_pal_2}} for secondary colors of the FU Berlin; 
\code{\link{fu_pal_3}} for a blue color gradient of the FU Berlin;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}}
}
\author{
\strong{unikn}, 2022-12-03.
}
\concept{contributed color palettes}
\keyword{datasets}
