% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_color.R
\name{get_col_names}
\alias{get_col_names}
\title{Get color names}
\usage{
get_col_names(col, custom_pals = all_pals)
}
\arguments{
\item{col}{A vector of colors.}

\item{custom_pals}{A vector of color palettes.}
}
\value{
A vector of color names.
}
\description{
\code{get_col_names} gets color names from color palettes.
}
\concept{utility functions}
