% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{grep_col}
\alias{grep_col}
\title{Get a vector of colors whose names match a regular expression.}
\usage{
grep_col(pattern, x = colors(), ignore_case = TRUE)
}
\arguments{
\item{pattern}{A regular expression 
(specified as a string/character object).}

\item{x}{A vector of R color names or a data frame of named colors
(i.e., whose names can be searched). 
Default: \code{x = colors()}.}

\item{ignore_case}{Should the case of pattern be ignored 
(passed to \code{ignore.case} of the \code{grep} function)?   
Default: \code{ignore_case = TRUE}.}
}
\description{
\code{grep_col} returns a vector of colors whose names match a regular expression (regex).
}
\details{
By default, the base R vector of named colors (i.e., \code{colors()}) is searched 
for names matching a \code{pattern} (which can be a simple string or regular expression). 

If \code{x} (i.e., the object to be searched) is provided, 
it is must be a vector of color names or a data frame of named color objects 
(e.g., a color palette).
}
\examples{
grep_col("cyan")

# With regular expressions:
grep_col("gr(a|e)y")
grep_col("^gr(a|e)y")
grep_col("^gr(a|e)y$")

# With other color objects (df as x):
grep_col("blau", x = pal_unikn)
grep_col("SEE", x = pal_unikn_pref)

# Example applications:
seecol(grep_col("cyan"))
seecol(grep_col("white"), col_bg = "lightblue2")
seecol(grep_col("SEE", pal_unikn))
seecol(grep_col("blau", pal_unikn_pref))

}
\seealso{
\code{\link{defpal}} to define color palettes; 
\code{\link{seepal}} to plot color palettes;  
\code{\link{usecol}} to use a color palette.

Other color functions: 
\code{\link{newpal}()},
\code{\link{seecol}()},
\code{\link{usecol}()}
}
\concept{color functions}
