% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text.R
\name{highlight}
\alias{highlight}
\title{\code{highlight} places 1 or more text strings (of a character vector \code{lbls}) 
onto a plot and places a colored box behind
each label to highlight it (i.e., make it stand out from the background).}
\usage{
highlight(x, y, lbls = NA, col_lbl = "black",
  col_bg = unlist(seeblau), cex = 2, font = 2)
}
\arguments{
\item{x}{A numeric vector of x-coordinates at which the 
text labels in \code{lbls} should be written. 
If the length of \code{x} and \code{y} differs, the shorter one is recycled.}

\item{y}{A numeric vector of y-coordinates at which the 
text labels in \code{lbls} should be written.}

\item{lbls}{A character vector specifying the text labels 
to be written.}

\item{col_lbl}{The color(s) of the text label(s). 
Default: \code{col_lbl = "black"}.}

\item{col_bg}{The color(s) to highlight or fill the rectangle(s) with. 
Default: \code{col_bg = NA} (transparent rectangles).}

\item{cex}{A numeric character expansion factor, 
multiplied by \code{par("cex")} to yield the character size. 
Default: \code{cex = 2}.}

\item{font}{The font to be used. 
Default: \code{font = 2} (i.e., bold).}
}
\description{
\code{highlight} uses the base graphics system \code{graphics::}.
}
\examples{
## Example 1: Simple highlights
plot(x = 0, y = 0, type = "n", xlim = c(0, 1), ylim = c(0, 1), xlab = "", ylab = "")

highlight(x = 0, y = .9, lbls = "Something to notice")
highlight(x = 0, y = c(.7, .6), 
          lbls = c("Highlighting text is simple", "but strikingly effective"), 
          cex = 1.5, col_bg = c(pal_seeblau[[2]], pal_seeblau[[1]]))
highlight(x = .5, y = c(.4, .3), lbls = c("It is also flexible", "but should be handled with care"), cex = 1.2, 
          col_lbl = c("white", "black"), col_bg = c(pal_seeblau[[5]], "gold"))

## Example 2: Messy plot
n <- 20
set.seed(1)
plot(x = runif(n), y = runif(n), type = "p", pch = 16, cex = 20, col = grey(0, .20), axes = F, xlab = "", ylab = "")

# Only 1 label:
highlight(x = .05, y = .90, lbls = "What a messy plot")

# 2 labels at once:
highlight(x = c(.30, .60), y = c(.20, .40),
          lbls = c("Note something here", "Some highlighting here"),
          col_bg = c(pal_seeblau[[2]], "gold"), cex = 1.2)
          
}
\seealso{
\code{\link{plot_box}} to plot text in a colored box

Other text functions: \code{\link{box_text}},
  \code{\link{plot_box}}
}
\concept{text functions}
