% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{hu_pal_1}
\alias{hu_pal_1}
\title{Primary colors of the Humboldt University Berlin}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on 
\url{https://www.hu-berlin.de/de/service/design/basiselemente/farbe}.
}
\usage{
hu_pal_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{hu_pal_1} provides the three primary colors 
of the \href{https://www.hu-berlin.de/en}{Humboldt University Berlin}, Germany.
}
\details{
The color \code{"HU blue"} is defined as Pantone 294 and only approximated by RGB 0/55/108.
Other approximations include CMYK 100/60/0/20, RGB 0/51/102 (reduced web scale), and RAL 270 3040.

See the \code{\link{ac}} and \code{\link{usecol}} functions for creating color gradients.
}
\examples{
hu_pal_1
seecol(hu_pal_1, n = 5, main = "5 shades of HU Berlin")
demopal(hu_pal_1, type = 1, main = "HU Berlin colors")

}
\seealso{
\code{\link{hu_pal_2}} for secondary colors of the HU Berlin;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}}
}
\author{
\strong{unikn}, 2022-11-30.
}
\concept{contributed color palettes}
\keyword{datasets}
