% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{inst}
\alias{inst}
\title{of \href{https://www.uni-kiel.de/en/}{Kiel University}, Germany.}
\format{
An object of class \code{character} of length 50.
}
\source{
Color definitions are based on 
\url{https://www.presse.uni-kiel.de/de/erscheinungsbild/farben}.
}
\usage{
inst
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
These colors may be mixed with darker shades of \code{"grey"} (or \code{"black"}). 
See examples and the \code{\link{ac}} and \code{\link{usecol}} functions 
for creating color gradients.
}
\details{
The full name of \href{https://www.uni-kiel.de/en/}{Kiel University} 
is Christian-Albrechts-Universität (CAU) zu Kiel.
}
\author{
\strong{unikn}, 2023-01-02.
}
\keyword{datasets}
