% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{karpfenblau}
\alias{karpfenblau}
\title{uni.kn color karpfenblau.}
\format{An object of class \code{data.frame} with 1 rows and 1 columns.}
\usage{
karpfenblau
}
\description{
\code{karpfenblau} provides the preferred color of \code{\link{pal_karpfenblau}} 
as an atomic data frame and is defined as 
\code{\link{pal_karpfenblau}[4]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
karpfenblau  # "karpfenblau4" OR "#3E5496"
all.equal(karpfenblau, pal_karpfenblau[4])  # TRUE

}
\seealso{
\code{\link{pal_karpfenblau}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} for \code{n} dedicated colors of a known color palette.

Other colors: \code{\link{grau}}, \code{\link{peach}},
  \code{\link{petrol}}, \code{\link{seeblau}},
  \code{\link{seegruen}}
}
\concept{colors}
\keyword{datasets}
