% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text_calls.R
\name{line}
\alias{line}
\title{Plot underlined text elements.}
\usage{
line(lbls, x = 0, y = 0.55, y_layout = "even", col = "black",
  col_bg = Seeblau, cex = 1.5, font = 1, new_plot = "none")
}
\arguments{
\item{lbls}{A character vector specifying the text labels 
to be written.}

\item{x}{A numeric vector of x-coordinates at which the 
text labels in \code{lbls} should be written. 
If the lengths of \code{x} and \code{y} differ, 
the shorter one is recycled. 
Default: \code{x = 0}.}

\item{y}{A numeric vector of y-coordinates at which the 
text labels in \code{lbls} should be written. 
If the lengths of \code{x} and \code{y} differ, 
the shorter one is recycled. 
Default: \code{y = .55}.}

\item{y_layout}{A numeric value or vector for the vertical 
spacing of labels in \code{lbls}. 
2 special values are 
\code{"even"} (i.e., even distribution of labels across available y-space) and 
\code{"flush"} (i.e., no space between adjacent labels, or \code{y_layout = 0}). 
Default: \code{y_layout = "even"}.}

\item{col}{The color(s) of the text label(s). 
Default: \code{col_lbl = "black"}.}

\item{col_bg}{The color(s) of the line (under the text labels 
of \code{lbls}). 
Default: \code{col_bg = Seeblau}.}

\item{cex}{Numeric character expansion factor(s), 
multiplied by \code{par("cex")} to yield the character size(s). 
Default: \code{cex = 1.5}.}

\item{font}{The font type(s) to be used. 
Default: \code{font = 1} (i.e., plain text).}

\item{new_plot}{Boolean: Should a new plot be generated?
Set to \code{"blank"} or \code{"slide"} to create a new plot. 
Default: \code{new_plot = "none"} (i.e., add to an existing plot).}
}
\description{
\code{line} plots 1 or more text strings (provided as a character vector \code{lbls}) 
to an (existing or new) plot and places a colored line beneath  
each label to underline it.
}
\details{
The positions of the text elements in \code{lbls} can be specified by 
providing their coordinates (as \code{x} and \code{y} arguments) or 
by providing an initial position and an \code{y_layout} (see below).

Text formatting parameters (like \code{col}, \code{col_bg}, \code{cex}, \code{font})         
are recycled to match \code{length(lbls)}.

\code{line} uses the base graphics system \code{graphics::}.
}
\examples{
lbl_line <- c("This is neat, true, and terribly important.")
line(lbls = lbl_line, new_plot = "blank")                # create a new plot
line(lbls = "(which is why we underline it).", y = .40)  # add to plot
                        
}
\seealso{
\code{\link{slide}} and \code{\link{xbox}} to create simple plots (without text).

Other text functions: \code{\link{box_text}},
  \code{\link{heading}}, \code{\link{mark}},
  \code{\link{plot_box_txt}}, \code{\link{post}},
  \code{\link{url_unikn}}
}
\concept{text functions}
