% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{lmu_pal_2}
\alias{lmu_pal_2}
\title{Secondary colors of the LMU Munich}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the 
\href{https://www.lmu.de/de/die-lmu/struktur/zentrale-universitaetsverwaltung/kommunikation-und-presse/lmu-brand-guide/designgrundsaetze/farben/index.html}{LMU brand guide}.
}
\usage{
lmu_pal_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{lmu_pal_2} provides four secondary colors 
of the \href{https://www.lmu.de/en/index.html}{Ludwig-Maximilians-Universität München}, Germany.
}
\details{
The secondary colors of \code{lmu_pal_2} are used to support the primary colors of \code{\link{lmu_pal_1}}.
}
\examples{
lmu_pal_2
seecol(lmu_pal_2, main = "The secondary/grey colors of LMU Munich")
demopal(lmu_pal_2, type = 5, main = "Secondary/grey colors of LMU Munich")

# Combining primary and secondary LMU colors:
lmu_pal_bipolor <- c(lmu_pal_1[-2], rev(lmu_pal_2), lmu_pal_1[2])
demopal(lmu_pal_bipolor, type = "mosaic", main = "Bipolar colors of LMU")

lmu_pal_linear <- c(lmu_pal_1[-3], lmu_pal_2, lmu_pal_1[3])
demopal(lmu_pal_linear, type = "polygon", seed = 2, main = "Linear colors of LMU")

}
\seealso{
\code{\link{lmu_pal_1}} for primary colors of the LMU Munich; 
\code{\link{lmu_pal_3}} for accent colors of the LMU Munich;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_3}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}}
}
\author{
\strong{unikn}, 2022-12-02.
}
\concept{contributed color palettes}
\keyword{datasets}
