% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text_calls.R
\name{mark}
\alias{mark}
\title{\code{mark} plots 1 or more text strings (provided as a character vector \code{lbls}) 
to an (existing or new) plot and places a colored box behind
each label to mark it (i.e., highlight or make it stand out from the background).}
\usage{
mark(lbls, x = 0, y = 0.55, y_layout = "even", col = "black",
  col_bg = Seeblau, cex = 2, font = 2, new_plot = "none")
}
\arguments{
\item{lbls}{A character vector specifying the text labels 
to be written.}

\item{x}{A numeric vector of x-coordinates at which the 
text labels in \code{lbls} should be written. 
If the lengths of \code{x} and \code{y} differ, 
the shorter one is recycled. 
Default: \code{x = 0}.}

\item{y}{A numeric vector of y-coordinates at which the 
text labels in \code{lbls} should be written. 
If the lengths of \code{x} and \code{y} differ, 
the shorter one is recycled. 
Default: \code{y = .55}.}

\item{y_layout}{A numeric value or vector for the vertical 
spacing of labels in \code{lbls}. 
2 special values are 
\code{"even"} (i.e., even distribution of labels across available y-space) and 
\code{"flush"} (i.e., no space between adjacent labels, or \code{y_layout = 0}). 
Default: \code{y_layout = "even"}.}

\item{col}{The color(s) of the text label(s). 
Default: \code{col_lbl = "black"}.}

\item{col_bg}{The color(s) to highlight or fill the rectangle(s) with. 
Default: \code{col_bg = Seeblau}.}

\item{cex}{Numeric character expansion factor(s), 
multiplied by \code{par("cex")} to yield the character size(s). 
Default: \code{cex = 2}.}

\item{font}{The font type(s) to be used. 
Default: \code{font = 2} (i.e., bold).}

\item{new_plot}{Boolean: Should a new plot be generated?
Set to \code{"blank"} or \code{"slide"} to create a new plot. 
Default: \code{new_plot = "none"} (i.e., add to an existing plot).}
}
\description{
Text formatting parameters (like \code{col}, \code{col_bg}, \code{cex}, \code{font})         
are recycled to match \code{length(lbls)}.
}
\details{
\code{mark} uses the base graphics system \code{graphics::}.
}
\examples{
mark(lbls = "Just testing.", new_plot = "blank")  # create a new plot
mark(lbls = "More testing here.", y = .45, col_bg = pal_pinky[[2]])  # add to plot
                        
}
\seealso{
\code{\link{slide}} and \code{\link{xbox}} to create simple plots (without text).

Other text functions: \code{\link{box_text}},
  \code{\link{heading}}, \code{\link{line}},
  \code{\link{plot_box_txt}}, \code{\link{post}},
  \code{\link{url_unikn}}
}
\concept{text functions}
