% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_color_fun.R
\name{mixcol}
\alias{mixcol}
\title{Provide a color gradient scale.}
\usage{
mixcol(pal = pal_unikn, n = length(pal), ...)
}
\arguments{
\item{pal}{A color palette (as a vector of colors or color palettes). 
Default: \code{pal = \link{pal_unikn}}.}

\item{n}{Number of desired colors. 
Default: \code{n = length(col)}.}

\item{...}{Additional parameters 
(passed to \code{grDevices::colorRampPalette}).}
}
\description{
\code{mixcol} provides a color gradient 
based on a vector of input colors or color palettes 
defined by \code{pal}.
}
\details{
\code{mixcol} uses \code{grDevices::colorRampPalette} to  
provide a function that takes an integer argument 
(the required number of colors) to  
return a character vector of (rgb) colors 
interpolating color sequence provided by \code{col}.

For best results, consider combining existing color palettes 
with base color "white" (see Examples).
}
\examples{
## Creating color gradients: ------ 

# Gradients extending 1 color palette:
mixcol(n = 5)  # 5 colors of default scale pal_unikn
mixcol(pal_petrol, n = 10)  # 10 colors of pal_petrol

# Gradients over 2+ colors:
mixcol(c(Seeblau, Peach), n = 10)
mixcol(c(Signal, Petrol), n = 10)  
mixcol(c(Seeblau, "white", Pinky), n = 10)
mixcol(c(Karpfenblau, Seeblau, "gold"), n = 10)  # "gold" shines brighter than signal

# Gradients over 2+ color palettes:
mixcol(c(pal_seeblau, pal_peach), n = 10)
mixcol(c(pal_seeblau, pal_karpfenblau), n = 10)
mixcol(c(pal_seeblau, pal_grau, pal_petrol), n = 10)

# Gradients over 3+ color palettes and colors:
mixcol(c(rev(pal_seeblau), "white", pal_pinky), n = 11)
mixcol(c(rev(pal_seeblau), "white", pal_petrol), n = 11)
mixcol(c(rev(pal_karpfenblau), "white", pal_bordeaux), n = 11) 

## Creating and plotting color gradients: ------ 

# (1) Extending color palettes: ----
seecol(mixcol(n = 20))
seecol(mixcol(pal_seeblau, n = 10))
seecol(mixcol(pal_bordeaux, n = 10))

# (2) Combining colors or palettes: ---- 

# Combining 2+ colors to create new palettes:
seecol(mixcol(c(Seeblau, Signal), n = 10))
seecol(mixcol(c(Signal, Petrol), n = 10))
seecol(mixcol(c(Seeblau, "white", Pinky), n = 10))
seecol(mixcol(c(Karpfenblau, Seeblau, "white"), n = 10))
seecol(mixcol(c(Bordeaux, "white", Petrol), n = 10))
seecol(mixcol(c(Seeblau, "white", Petrol), n = 10))
seecol(mixcol(c(Karpfenblau, Seeblau, "gold"), n = 10))  # "gold" shines brighter than signal
  
# Combining 2+ color palettes into new palettes:
seecol(mixcol(c(rev(pal_seeblau), pal_petrol), n = 10))
seecol(mixcol(c(rev(pal_seeblau), pal_peach), n = 10))
seecol(mixcol(c(rev(pal_karpfenblau), pal_bordeaux), n = 10))
seecol(mixcol(c(rev(pal_seegruen), pal_pinky), n = 10))
seecol(mixcol(c(rev(pal_petrol), pal_bordeaux), n = 10))

#' # Combining 2 palettes and mid-color "white" to new palettes:
seecol(mixcol(c(rev(pal_seeblau), "white", pal_pinky), n = 11))
seecol(mixcol(c(rev(pal_seeblau), "white", pal_petrol), n = 11))
seecol(mixcol(c(rev(pal_karpfenblau), "white", pal_bordeaux), n = 11))
seecol(mixcol(c(rev(pal_seegruen), "white", pal_peach), n = 11))
seecol(mixcol(c(rev(pal_petrol), "white", pal_bordeaux), n = 11))

}
\seealso{
\code{\link{seecol}} to plot color palettes; 
\code{\link{pal_which}} to get specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_peach}}, \code{\link{pal_petrol}},
  \code{\link{pal_pinky}}, \code{\link{pal_seeblau}},
  \code{\link{pal_seegruen}}, \code{\link{pal_signal}},
  \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{plot_pal}}

Other color functions: \code{\link{pal_which}},
  \code{\link{plot_pal}}, \code{\link{seecol}},
  \code{\link{use_pal_n}}, \code{\link{usecol}}
}
\concept{color functions}
\concept{color palettes}
