% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{newpal}
\alias{newpal}
\alias{defpal}
\title{Define new color palettes.}
\usage{
newpal(col, names = NA, as_df = FALSE)
}
\arguments{
\item{col}{A required vector of colors 
(specified by their R color names or hex codes).}

\item{names}{A character vector of names 
Default: \code{names = NA}, yielding numeric names.}

\item{as_df}{Should the new color palette be returned as 
a data frame (rather than a vector)? 
Default: \code{as_df = FALSE}.}
}
\description{
\code{newpal} allows defining new color palettes 
(as data frames).
}
\examples{
# Define a color palette:
newpal(col = c("black", "white"), names = c("dark", "bright"))

pal_mpg <- newpal(col = c("#007367", "white", "#D0D3D4"), 
                  names = c("mpg green", "white", "mpg grey")
                  )
seecol(pal_mpg) 

}
\seealso{
\code{\link{seepal}} to plot color palettes;  
\code{\link{usecol}} to use a color palette.

Other color functions: \code{\link{seecol}},
  \code{\link{usecol}}
}
\concept{color functions}
