% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pal_Bordeaux}
\alias{pal_Bordeaux}
\title{uni.kn Bordeaux color palette.}
\format{An object of class \code{data.frame} with 1 rows and 5 columns.}
\usage{
pal_Bordeaux
}
\description{
\code{pal_Bordeaux} provides an additional uni.kn color palette  
as a data frame containing 5 colors (shades of Bordeaux).
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
pal_Bordeaux
dim(pal_Bordeaux)  # 1 5
pal_Bordeaux[4]    # preferred (named) color "Bordeaux4"
pal_Bordeaux[[4]]  # preferred color "Bordeaux4" OR "#8E2043"

# Plotting palette:
plot_pal(pal_Bordeaux)

}
\seealso{
\code{\link{pal_peach}} and \code{\link{pal_pinky}} for alternative redish uni.kn color palettes;  
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{col_scale}} to extend color palettes (by creating gradients).

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_grau}}, \code{\link{pal_karpfenblau}},
  \code{\link{pal_n_sq}}, \code{\link{pal_n}},
  \code{\link{pal_peach}}, \code{\link{pal_petrol}},
  \code{\link{pal_pinky}}, \code{\link{pal_seeblau}},
  \code{\link{pal_seegruen}}, \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
