% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pal_grau}
\alias{pal_grau}
\title{uni.kn grau color palette.}
\format{An object of class \code{data.frame} with 1 rows and 5 columns.}
\usage{
pal_grau
}
\description{
\code{pal_grau} provides an additional uni.kn color palette  
as a data frame containing 5 colors (shades of grey).
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
pal_grau
dim(pal_grau)  # 1 5
pal_grau[3]    # preferred (named) color "grau3"
pal_grau[[3]]  # preferred color "grau3" OR "#9AA0A7"

}
\seealso{
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_n}} for \code{n} dedicated colors of a known color palette.

Other color palettes: \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_seeblau}},
  \code{\link{pal_seegruen}}, \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
