% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{pal_n_sq}
\alias{pal_n_sq}
\title{Get n^2 dedicated colors of a color palette.}
\usage{
pal_n_sq(n = "all", pal = pal_unikn)
}
\arguments{
\item{n}{A number specifying the desired number colors of pal (as a number) 
or the character string \code{"all"} (to get all colors of \code{pal}). 
Default: \code{n = "all"}.}

\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{pal_n_sq} returns \code{n^2} dedicated colors of a color palette \code{pal} 
(up to a maximum of \code{n + 1} colors).
}
\details{
Note that \code{pal_n_sq} was created for \code{pal = \link{pal_unikn}} 
for small values of \code{n} (\code{n = 1, 2, 3}) and 
returns the 11 colors of \code{\link{pal_unikn_plus}} for any \code{n > 3}.

Use the more specialized function \code{\link{pal_n}} for choosing 
\code{n} dedicated colors of a known color palette.
}
\examples{
pal_n_sq(1) #  1 color: seeblau3
pal_n_sq(2) #  4 colors
pal_n_sq(3) #  9 colors (5: white)
pal_n_sq(4) # 11 colors of pal_unikn_plus (6: white)

}
\seealso{
\code{\link{plot_pal}} to plot color palettes; 
\code{\link{col_scale}} to create color scales (gradients);  
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{pal_which}} to get specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n}},
  \code{\link{pal_peach}}, \code{\link{pal_petrol}},
  \code{\link{pal_pinky}}, \code{\link{pal_seeblau}},
  \code{\link{pal_seegruen}}, \code{\link{pal_signal}},
  \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}},
  \code{\link{seecol}}
}
\concept{color palettes}
