% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pal_peach}
\alias{pal_peach}
\title{uni.kn peach color palette.}
\format{An object of class \code{data.frame} with 1 rows and 5 columns.}
\usage{
pal_peach
}
\description{
\code{pal_peach} provides an additional uni.kn color palette  
as a data frame containing 5 colors (shades of peach).
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
pal_peach
dim(pal_peach)  # 1 5
pal_peach[4]    # preferred (named) color "peach4" 
pal_peach[[4]]  # preferred color "peach4" OR "#FEA090"

# Plotting palette:
plot_pal(pal_peach)

}
\seealso{
\code{\link{pal_pinky}} and \code{\link{pal_Bordeaux}} for alternative redish uni.kn color palettes;  
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{col_scale}} to extend color palettes (by creating gradients).

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_Bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_petrol}},
  \code{\link{pal_pinky}}, \code{\link{pal_seeblau}},
  \code{\link{pal_seegruen}}, \code{\link{pal_signal}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
