% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\docType{data}
\name{pal_seeblau}
\alias{pal_seeblau}
\title{uni.kn seeblau color palette.}
\format{An object of class \code{data.frame} with 1 rows and 5 columns.}
\usage{
pal_seeblau
}
\description{
\code{pal_seeblau} provides an additional uni.kn color palette 
as a data frame containing 5 colors (shades of blue/seeblau).
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
pal_seeblau
dim(pal_seeblau)  # 1 5
pal_seeblau[3]    # preferred (named) color "seeblau3" 
pal_seeblau[[3]]  # preferred color "seeblau3" = "#59C7EB"

# Plotting palette:
plot_pal(pal_seeblau)

}
\seealso{
\code{\link{pal_karpfenblau}} for an alternative blue uni.kn color palette;  
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{col_scale}} to extend color palettes (by creating gradients).

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_Bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seegruen}}, \code{\link{pal_signal}},
  \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
