% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pal_unikn}
\alias{pal_unikn}
\title{uni.kn default color scale.}
\format{An object of class \code{data.frame} with 1 rows and 10 columns.}
\usage{
pal_unikn
}
\description{
\code{pal_unikn} provides an uni.kn color scale 
as a data frame containing 10 colors.
}
\details{
This is the default (web/sRGB) scale.

See https://www.uni-konstanz.de for details.
}
\examples{
pal_unikn
dim(pal_unikn)  # 1 10
pal_unikn[1]    # named color
pal_unikn[[1]]  # color

}
\seealso{
\code{\link{pal_unikn_ppt}} for a ppt version.

Other color scales: \code{\link{col_n_sq}},
  \code{\link{col_n}}, \code{\link{col_which}},
  \code{\link{pal_unikn_ppt}}, \code{\link{plot_pal}}
}
\concept{color scales}
\keyword{datasets}
