% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pal_unikn_plus}
\alias{pal_unikn_plus}
\title{uni.kn 11 colors in a color palette.}
\format{An object of class \code{data.frame} with 1 rows and 11 columns.}
\usage{
pal_unikn_plus
}
\description{
\code{pal_unikn_plus} combines the 5 blue colors 
from color palette \code{\link{pal_seeblau}} 
with the 6 non-blue colors of \code{\link{pal_unikn}} 
to a palette containing 11 color values.
}
\details{
Adding seeblau5 (i.e., \code{pal_seeblau[1]}) to 
the default color palette \code{\link{pal_unikn}} 
also puts \code{white} at the central (middle) 
position of a palette with 11 values:

\code{pal_unikn_plus[[6]]} is \code{white} or \code{"#FFFFFF"}.

This is useful when creating color gradients.

See \url{https://www.uni-konstanz.de} for details.
}
\examples{
pal_unikn_plus
dim(pal_unikn_plus)         # 1 11
pal_unikn_plus[1]           # seeblau5 (new)
pal_unikn_plus[[1]]         # darker blue color value: "#008ECE"
pal_unikn_plus["seeblau5"]  # seeblau5 color value by name

# Note:
pal_unikn_plus[6] # "white" or "#FFFFFF" as central of 11 colors

}
\seealso{
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_seeblau}} for the uni.kn seeblau color palette; 
\code{\link{pal_n}} for \code{n} dedicated colors of a known color palette.

Other color palettes: \code{\link{pal_grau}},
  \code{\link{pal_n_sq}}, \code{\link{pal_n}},
  \code{\link{pal_peach}}, \code{\link{pal_petrol}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
