% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pal_unikn_ppt}
\alias{pal_unikn_ppt}
\title{uni.kn secondary color palette (ppt version).}
\format{An object of class \code{data.frame} with 1 rows and 10 columns.}
\usage{
pal_unikn_ppt
}
\description{
\code{pal_unikn_ppt} provides an alternative uni.kn color palette  
as a data frame containing 10 colors.
}
\details{
This is a secondary (ppt) variant with more muted colors.

See \code{\link{pal_unikn}} for the primary/default (web/sRGB) scale 
and \url{https://www.uni-konstanz.de} for details.
}
\examples{
pal_unikn_ppt
dim(pal_unikn_ppt)  # 1 10
pal_unikn_ppt[1]    # named color
pal_unikn_ppt[[1]]  # color

}
\seealso{
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_n}} for \code{n} dedicated colors of a known color palette.

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_grau}}, \code{\link{pal_karpfenblau}},
  \code{\link{pal_n_sq}}, \code{\link{pal_n}},
  \code{\link{pal_peach}}, \code{\link{pal_petrol}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
