% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{Petrol}
\alias{Petrol}
\title{uni.kn color Petrol.}
\format{An object of class \code{character} of length 1.}
\usage{
Petrol
}
\description{
\code{Petrol} provides the preferred color of \code{\link{pal_petrol}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_petrol}[[4]]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
Petrol  # HEX character "#077187" (as value)
all.equal(Petrol, pal_petrol[[4]])  # TRUE (same HEX values)

}
\seealso{
\code{\link{pal_unikn}} for the unikn default color palette with all 5 colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_petrol}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: \code{\link{Bordeaux}},
  \code{\link{Grau}}, \code{\link{Karpfenblau}},
  \code{\link{Peach}}, \code{\link{Pinky}},
  \code{\link{Seeblau}}, \code{\link{Seegruen}},
  \code{\link{Signal}}
}
\concept{preferred colors}
\keyword{datasets}
