% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{Pinky}
\alias{Pinky}
\title{uni.kn color pinky (as df).}
\format{An object of class \code{character} of length 1.}
\usage{
Pinky
}
\description{
\code{pinky} provides the preferred color of \code{\link{pal_pinky}} 
(as an atomic data frame) and is defined as 
\code{\link{pal_pinky}[4]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
# pinky  # name "pinky4" and color value "#E0607E" (as df)
# all.equal(pinky, pal_pinky[4])  # TRUE (both df)

Pinky  # HEX character "#E0607E" (as value)
all.equal(Pinky, pal_pinky[[4]])  # TRUE (same HEX values)

}
\seealso{
\code{\link{pal_pinky}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{seepal}} to show and use color palettes. 
uni.kn color Pinky.   

\code{Pinky} provides the preferred color of \code{\link{pal_pinky}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_pinky}[[4]]}. 

See \url{https://www.uni-konstanz.de} for details.

\code{\link{pal_pinky}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{seepal}} to show and use color palettes.

Other named colors (as df): \code{\link{Bordeaux}},
  \code{\link{Grau}}, \code{\link{Karpfenblau}},
  \code{\link{Peach}}, \code{\link{Petrol}},
  \code{\link{Seeblau}}, \code{\link{Seegruen}},
  \code{\link{Signal}}

Other preferred colors: \code{\link{Bordeaux}},
  \code{\link{Grau}}, \code{\link{Karpfenblau}},
  \code{\link{Peach}}, \code{\link{Petrol}},
  \code{\link{Seeblau}}, \code{\link{Seegruen}},
  \code{\link{Signal}}
}
\concept{named colors (as df)}
\concept{preferred colors}
\keyword{datasets}
