% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box_txt}
\alias{plot_box_txt}
\title{Plot a colored box (with top x and text).}
\usage{
plot_box_txt(lbls = "", col_lbl = "white", col_bg = Seeblau,
  cex = 2, font = 2, lbl_x = 0.02, lbl_y = 0.65)
}
\arguments{
\item{lbls}{A character vector specifying the text label(s)
to be written. If \code{lbls} contains multiple character strings, 
each one is placed on a new line. 
Default: \code{lbls = NA}.}

\item{col_lbl}{The color(s) of the text label(s).
Default: \code{col_lbl = "white"}.}

\item{col_bg}{The color(s) to fill the box with. 
Default: \code{col_bg = seeblau}.}

\item{cex}{A numeric character expansion factor, which is
multiplied by \code{par("cex")} to yield the final character size. 
Default: \code{cex = 2}.}

\item{font}{The font to be used. 
Default: \code{font = 2} (bold).}

\item{lbl_x}{A fraction in \code{[0, 1]} that determines the 
x-value of the left edge of text in \code{lbls}. 
Default: \code{lbl_x = .02} (i.e., 2\% from the left border).}

\item{lbl_y}{A fraction in \code{[0, 1]} that determines the 
y-value of the top line of text in \code{lbls}.  
To safe-guard the space to the left of the top cross, 
a maximum value of 1 corresponds to 80\% of the box height.) 
Default: \code{lbl_y = .65} (i.e., 65\% x 80\% = 52\% of box height).}
}
\description{
\code{plot_box_txt} generates a box with a cross (x) in its top-right corner 
and places 1 or more text strings (of a character vector \code{lbls}) 
in the box.
}
\details{
\code{plot_box_txt} uses the base graphics system \code{graphics::}.

This is an earlier version of the more powerful 
\code{plot_box} function.
}
\examples{
# plot_box_txt()  # default box (w/o text)

# plot_box_txt(lbls = "A heading appears here.")
# plot_box_txt(lbls = c("Some title", "The second line is longer", "A third short line"), cex = 2.4)

# plot_box_txt(lbls = "The darkest shade of seeblau.", col_bg = pal_seeblau[[5]], cex = 1.5)

## Note: As a vector of character strings for lbls is converted into separate lines of text,
##       the following examples yield identical results:
# plot_box_txt(lbls = c("1st line of text.", 
#                       "A 2nd longer line of text.", 
#                       "The 3rd line of text."))
# plot_box_txt(lbls = "1st line of text.\\nA 2nd longer line of text.\\nThe 3rd line of text.")

## Box logos:
# plot_box_txt(lbls = c("unikn::"), col_bg = Seeblau, cex = 2.5)
# plot_box_txt(lbls = "ToDo", cex = 4, col_bg = unlist(pal_bordeaux[4]))
# plot_box_txt(lbls = "R", col_bg = pal_seeblau[[5]], cex = 10, lbl_y = .7)

}
\seealso{
\code{\link{mark}} to mark text with a colored box.

Other text functions: \code{\link{box_text}},
  \code{\link{heading}}, \code{\link{line}},
  \code{\link{mark}}, \code{\link{post}}
}
\concept{text functions}
