% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plot_pal}
\alias{plot_pal}
\title{Plot a color palette.}
\usage{
plot_pal(pal = pal_unikn)
}
\arguments{
\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{plot_pal} plots (information on) a color palette \code{pal}.
}
\examples{
plot_pal()  # plots default color palette 

# RColorBrewer palettes: 
plot_pal(RColorBrewer::brewer.pal(n = 5, name = "Set1"))
plot_pal(RColorBrewer::brewer.pal(n = 5, name = "Blues"))

# yarrr palettes: 
plot_pal(yarrr::piratepal(palette = "appletv"))
plot_pal(yarrr::piratepal(palette = "google"))

}
\seealso{
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color scales: \code{\link{col_n_sq}},
  \code{\link{col_n}}, \code{\link{col_which}},
  \code{\link{pal_unikn_ppt}}, \code{\link{pal_unikn}}
}
\concept{color scales}
