% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{seeblau}
\alias{seeblau}
\title{uni.kn color seeblau.}
\format{An object of class \code{data.frame} with 1 rows and 1 columns.}
\usage{
seeblau
}
\description{
\code{seeblau} provides the preferred color of \code{\link{pal_seeblau}} 
as an atomic data frame and is defined as 
\code{\link{pal_seeblau}[3]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
seeblau  # "seeblau3" OR "#59C7EB"
all.equal(seeblau, pal_seeblau[3])  # TRUE

}
\seealso{
\code{\link{pal_seeblau}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette.

Other colors: \code{\link{Bordeaux}}, \code{\link{grau}},
  \code{\link{karpfenblau}}, \code{\link{peach}},
  \code{\link{petrol}}, \code{\link{pinky}},
  \code{\link{seegruen}}
}
\concept{colors}
\keyword{datasets}
