% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{seecol}
\alias{seecol}
\alias{seepal}
\title{Plot color palettes (to see their colors).}
\usage{
seecol(pal = "all", n = "all", hex = NULL, rgb = NULL,
  col_brd = NULL, grid = TRUE, ...)
}
\arguments{
\item{pal}{A color palette (as a list of colors) or 
the character string \code{"all"} (for showing all 
color palettes in the current environment). 
Default: \code{pal = "all"}.}

\item{n}{Number of colors to show or use. 
If \code{n} is lower or higher than the length of the current 
color palette \code{pal}, the color palette is reduced or extrapolated 
(using \code{grDevices::colorRampPalette}). 
Default: \code{n = "all"}.}

\item{hex}{Should HEX color values be shown?
Default: \code{hex = NULL} (i.e., show HEX color values 
when there is sufficient space to print them).}

\item{rgb}{Should RGB color values be shown?
Default: \code{rgb = NULL} (i.e., show RGB color values 
when there is sufficient space to print them).}

\item{col_brd}{Color of box borders (if shown). 
Default: \code{col_brd = NULL}.}

\item{grid}{Show grid?
Default: \code{grid = TRUE}.}

\item{...}{Other graphical parameters 
(passed to \code{plot_col}).}
}
\description{
\code{seecol} provides an interface to plotting (or "seeing") 
the colors of a palette or comparing multiple color palettes.
}
\details{
\code{seecol} has 2 main modes, based on the contents of its \code{pal} argument:

\enumerate{

\item if \code{pal = "all"} (or a list of multiple color palettes):

Plot visual vectors of all current color palettes for comparing them.

\item if \code{pal} is set to a specific color palette (or a vector of multiple colors or color palettes):

Plot the current color palette and optional details on its colors.

}
}
\examples{
# See all color palettes: 
seecol()  # same as seecol(pal = "all") 

# See details of a color palette: 
seecol(pal_unikn_plus)  # see a specific color palette

# Combining colors or color palettes: 
seecol(c("black", "firebrick", "gold"))
seecol(c(rev(pal_seeblau), pal_seegruen))
seecol(c(rev(pal_seeblau), "white", pal_pinky))

# Using n to reduce or extend color palettes:
seecol(n =  3)  
seecol(n = 12)

seecol(pal_unikn, n = 5)
seecol(pal_seeblau, n = 10)

# Combining and extending color palettes: 
seecol(c(rev(pal_seeblau), "white", pal_bordeaux), n = 17)

}
\seealso{
\code{\link{usecol}} to use a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color functions: \code{\link{usecol}}
}
\concept{color functions}
