% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{Seegruen}
\alias{Seegruen}
\title{uni.kn color seegruen (as df).}
\format{An object of class \code{character} of length 1.}
\usage{
Seegruen
}
\description{
\code{seegruen} provides the preferred color of \code{\link{pal_seegruen}} 
(as an atomic data frame) and is defined as 
\code{\link{pal_seegruen}[4]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
# seegruen  # name "seegruen4" and color value "#0A9086" (as df)
# all.equal(seegruen, pal_seegruen[4])  # TRUE (both df)

Seegruen  # HEX character "#0A9086" (as value)
all.equal(Seegruen, pal_seegruen[[4]])  # TRUE (same HEX values)

}
\seealso{
\code{\link{pal_seegruen}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{seepal}} to show and use color palettes. 
uni.kn color Seegruen. 

\code{Seegruen} provides the preferred color of \code{\link{pal_seegruen}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_seegruen}[[4]]}. 

See \url{https://www.uni-konstanz.de} for details.

\code{\link{pal_seegruen}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{seepal}} to show and use color palettes.

Other named colors (as df): \code{\link{Bordeaux}},
  \code{\link{Grau}}, \code{\link{Karpfenblau}},
  \code{\link{Peach}}, \code{\link{Petrol}},
  \code{\link{Pinky}}, \code{\link{Seeblau}},
  \code{\link{Signal}}

Other preferred colors: \code{\link{Bordeaux}},
  \code{\link{Grau}}, \code{\link{Karpfenblau}},
  \code{\link{Peach}}, \code{\link{Petrol}},
  \code{\link{Pinky}}, \code{\link{Seeblau}},
  \code{\link{Signal}}
}
\concept{named colors (as df)}
\concept{preferred colors}
\keyword{datasets}
