% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pal.R
\name{seepal}
\alias{seepal}
\title{Show and use color palettes.}
\usage{
seepal(pal = "all", n = "all", hex = NULL, rgb = NULL,
  col_brd = NULL, ...)
}
\arguments{
\item{pal}{A color palette (as a data frame) or 
the character string \code{"all"} (to show all 
color palettes in the current environment). 
Default: \code{pal = "all"}.}

\item{n}{Number of colors to show or use.
Default: \code{n = "all"}.}

\item{hex}{Should HEX color values be shown?
Default: \code{hex = NULL} (i.e., show HEX color values 
when there is sufficient space to print them).}

\item{rgb}{Should RGB color values be shown?
Default: \code{rgb = NULL} (i.e., show RGB color values 
when there is sufficient space to print them).}

\item{col_brd}{Color of box borders (if shown).}

\item{...}{Other graphical parameters 
(passed to \code{plot_col}).}
}
\description{
\code{seepal} provides an interface to color palettes.
}
\examples{
seepal()  # shows all color palettes

}
\seealso{
\code{\link{col_scale}} to extend color palettes (by creating gradients);  
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{pal_which}} to get specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
