% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{shades_of}
\alias{shades_of}
\title{Get n shades of a color.}
\usage{
shades_of(n = 5, col_1 = "black", col_n = "white", alpha = NA)
}
\arguments{
\item{n}{Number of desired colors.
Default: \code{n = 5}.}

\item{col_1}{Initial color.
Default: \code{col_1 = "black"}.}

\item{col_n}{Final (n-th) color.
Default: \code{col_n = "white"}.}

\item{alpha}{A factor modifying the opacity alpha 
(as in \code{\link{adjustcolor}}) to a value in \code{[0, 1]}. 
Default: \code{alpha = NA} (i.e., no modification of opacity).}
}
\description{
\code{shades_of} returns a vector of \code{n} colors that are 
shades of an initial color \code{col_1}.
}
\details{
By default, the colors range from the initial color 
\code{col_1 = "black"} to \code{col_n = "white"}, 
but specifying different initial and final colors yields 
other color ranges. 

\code{shades_of} is only a convenient wrapper function for 
a common \code{\link{usecol}} command. 
However, \code{\link{usecol}} allows defining more 
complex color gradients (e.g., by specifying more than two colors).
}
\examples{
grey50 <- shades_of(50, col_1 = 'grey1')
seecol(grey50, title = "50 shades of grey1")

blue_black <- shades_of(5, Seeblau, col_n = "black")
seecol(blue_black, title = "5 shades from Seeblau to black")

wine_white <- shades_of(6, Bordeaux, alpha = 1/2)
seecol(wine_white, col_brd = "black", lwd_brd = .5, 
       title = "Shades of semi-transparent Bordeaux")
  
}
\seealso{
\code{\link{seepal}} to plot color palettes;  
\code{\link{usecol}} to use a color palette; 
\code{\link{defpal}} to define new color palettes; 
\code{\link{grepal}} for finding named colors.

Other color functions: 
\code{\link{grepal}()},
\code{\link{newpal}()},
\code{\link{seecol}()},
\code{\link{usecol}()}
}
\concept{color functions}
