% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{simcol}
\alias{simcol}
\title{Find similar colors.}
\usage{
simcol(col_target, col_candidates = colors(), tol = c(25, 50, 75))
}
\arguments{
\item{col_target}{A (required) target color.}

\item{col_candidates}{Palette of color candidates. 
Default: \code{col_candidates = colors()}.}

\item{tol}{Numeric tolerance value(s) (must be 
either 1 or 3 numeric values, in RGB range from 0 to 255).
Default: \code{tol = c(25, 50, 75)}.}
}
\value{
A named vector of colors or color values.
}
\description{
\code{simcol} finds and shows colors 
from a palette of color candidates \code{col_candidates} that are similar 
to some target color \code{col_target}.
}
\details{
\code{simcol} returns a vector of (named) colors or color values, 
but also uses \code{\link{seecol}} to visualize its results.

Color similarity is defined in terms of the distance between colors'  
RGB values, within some numeric tolerance threshold(s) provided by \code{tol}. 
Higher \code{tol} values correspond to more permissive similarity judgments.  

If \code{tol} is a scalar, all three RGB values of \code{col_candidates}
must be within the corresponding values of \code{col_target} 
to be considered 'similar'.
If \code{tol} contains three values, the three RGB dimension are compared 
in order of the dimensions' rank in \code{col_target} 
(i.e., the primary dimension must be within \code{tol[1]}, etc.). 
Thus, providing three \code{tol} values allows for 
more fine-grained similarity matching.
}
\examples{
# Basic uses:
simcol(col_target = "red")
simcol(Seeblau, tol = 30)
simcol("gold", tol = c(20, 30, 40))
simcol("blue", col_candidates = pal_unikn_pref, tol = 120)

# More fine-grained color matching:
simcol(Seeblau, tol = 40)  # = simcol(Seeblau, tol = c(40, 40, 40))
simcol(Seeblau, tol = c(10, 30, 60))

}
\seealso{
\code{\link{seecol}} for plotting/seeing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors; 
\code{\link{shades_of}} to defining shades of a given color; 
\code{\link{ac}} for adjusting color transparency.

Other color functions: 
\code{\link{ac}()},
\code{\link{grepal}()},
\code{\link{newpal}()},
\code{\link{seecol}()},
\code{\link{shades_of}()},
\code{\link{usecol}()}
}
\concept{color functions}
