% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_bwkn}
\alias{theme_bwkn}
\title{An alternative bw-theme (for ggplot2)}
\usage{
theme_bwkn(
  col_title = grey(0.1, 1),
  col_line = grey(0.4, 1),
  base_size = 12,
  base_family = "sans",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{col_title}{Color of text elements 
used for plot title, axis titles, and legend (optional).
Default: \code{col_title = grey(.10, 1)} (i.e., dark grey).  
Consider using \code{col_title = unikn::pal_seeblau[[4]]}.}

\item{col_line}{Color of line elements 
used for axis lines, text and ticks, plot caption and subtitle, and panel border.
Default: \code{col_title = grey(.40, 1)} (i.e., medium grey).}

\item{base_size}{Base font size (optional, numeric). 
Default: \code{base_size = 12}.}

\item{base_family}{Base font family (optional, character). 
Default: \code{base_family = "sans"}. 
Options include \code{"mono"}, \code{"sans"} (default), and "serif".}

\item{base_line_size}{Base line size (optional, numeric). 
Default: \code{base_line_size = base_size/22}.}

\item{base_rect_size}{Base rectangle size (optional, numeric). 
Default: \code{base_rect_size = base_size/22}.}
}
\description{
\code{theme_bwkn} provides an alternative bw theme 
for use in \bold{ggplot2} and \bold{unikn} contexts.
}
\details{
\code{theme_bwkn} is light and no-nonsense, 
and based on \code{theme_bw}. 

This theme works well for most visualizations 
that require grid and plot panel boundaries.
}
\examples{

\donttest{
# Plotting iris dataset (using ggplot2, theme_grau, and unikn colors):
  
library('ggplot2')  # theme_bwkn requires ggplot2 
   
ggplot(datasets::iris) +
  geom_jitter(aes(x = Sepal.Length, y = Sepal.Width, color = Species), size = 3, alpha = 2/3) +
  facet_wrap(~Species) +
  scale_color_manual(values = usecol(pal = c(Pinky, Seeblau, Seegruen))) +
  labs(tag = "C",
       title = "Iris sepals by species (using theme_bwkn)",
       caption = "Data from datasets::iris") + 
  coord_fixed(ratio = 3/2) + 
  theme_bwkn()

}

}
\seealso{
\code{\link{theme_unikn}} for default theme;  
\code{\link{theme_grau}} for a grey theme; 
\code{\link{theme_minikn}} for a minimal theme.

Other plot functions: 
\code{\link{slide}()},
\code{\link{theme_grau}()},
\code{\link{theme_minikn}()},
\code{\link{theme_unikn}()},
\code{\link{xbox}()}
}
\concept{plot functions}
