% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_unikn}
\alias{theme_unikn}
\title{Basic unikn theme for ggplot2.}
\usage{
theme_unikn(col_title = unikn::pal_seeblau[[4]], base_size = 11,
  base_family = "", base_line_size = base_size/20,
  base_rect_size = base_size/20)
}
\arguments{
\item{col_title}{Color of title (text) elements (optional, numeric).
Default: \code{col_title = "black"}.  
Consider using \code{col_title = unikn::pal_seeblau[[4]]} 
in combination with black or grey data points.}

\item{base_size}{Base font size (optional, numeric). 
Default: \code{base_size = 11}.}

\item{base_family}{Base font family (optional, character). 
Default: \code{base_family = ""}.}

\item{base_line_size}{Base line size (optional, numeric). 
Default: \code{base_line_size = base_size/20}.}

\item{base_rect_size}{Base rectangle size (optional, numeric). 
Default: \code{base_rect_size = base_size/20}.}
}
\description{
\code{theme_unikn} provides a basic unikn theme 
to use in ggplot2 commands.
}
\details{
The theme is lightweight and no-nonsense, but somewhat 
opinionated (e.g., in using mostly grey scales to 
allow emphasizing data points with color accents).
}
\seealso{
Other plot functions: \code{\link{slide}},
  \code{\link{xbox}}
}
\concept{plot functions}
