% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{uni_freiburg_blue}
\alias{uni_freiburg_blue}
\title{Blue colors of the University of Freiburg.}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on 
\url{http://portal.uni-freiburg.de/cmsforum/wsg/webstyleguide/farben}.
}
\usage{
uni_freiburg_blue
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_freiburg_blue} provides the blue colors 
of the \href{https://uni-freiburg.de/}{University of Freiburg}, Germany.
}
\examples{
uni_freiburg_blue
seecol(uni_freiburg_blue)  # view color palette
demopal(uni_freiburg_blue, type = 3, main = "Blue colors of the University of Freiburg")

}
\seealso{
\code{\link{uni_freiburg_br}} for the default colors of the University of Freiburg;  
\code{\link{uni_freiburg_info}} for the info colors of the University of Freiburg;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{eth_pal_black_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{mpg_pal}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}}
}
\author{
Hansjoerg Neth, 2022-10-20.
}
\concept{contributed color palettes}
\keyword{datasets}
