% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{uni_goettingen_3}
\alias{uni_goettingen_3}
\title{Departmental colors of the University of Göttingen}
\format{
An object of class \code{character} of length 13.
}
\source{
Color definitions are based on 
\url{https://www.uni-goettingen.de/de/589412.html}.
}
\usage{
uni_goettingen_3
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_goettingen_3} provides the 13 departmental colors 
(German "Fakultäten") of the \href{https://www.uni-goettingen.de/}{University of Göttingen}, Germany.
}
\details{
These colors are used in online and print materials of these departments.

\code{uni_goettingen_3} corrects an error in the color definitions 
of \url{https://www.uni-goettingen.de/de/589412.html}, 
where the color definition for "Theologie" used non-corresponding RGB and HEX values. 
We adopted the RGB value \code{68 37 61} and its corresponding HEX value \code{"#44253D"}.
}
\examples{
uni_goettingen_3
seecol(uni_goettingen_3)  # view color palette
demopal(uni_goettingen_3, type = 3, main = "Departmental colors of the University of Göttingen")

}
\seealso{
\code{\link{uni_goettingen_1}} and \code{\link{uni_goettingen_2}} for alternative colors of the University of Goettingen;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{eth_pal_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{mpg_pal}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}}
}
\author{
Hansjoerg Neth, 2022-11-20.
}
\concept{contributed color palettes}
\keyword{datasets}
