% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{uni_koeln_1}
\alias{uni_koeln_1}
\title{Primary colors of the University of Koeln (Cologne)}
\format{
An object of class \code{character} of length 7.
}
\source{
Color definitions are based on 
\url{https://kommunikation-marketing.uni-koeln.de/marketing/corporate_design/index_ger.html}.
}
\usage{
uni_koeln_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_koeln_1} provides the primary colors 
of the \href{https://www.uni-koeln.de/}{Universitaet zu Koeln}, Germany.
}
\details{
The first 6 colors are variants of the main color \code{blaugrau} 
(defined as R122.G147.B171, see \code{uni_koeln_1[4]}). 

The 7th color is the accent color \code{signal rot} 
(defined as Pantone Red 032 C or R175.G17.B29, see \code{uni_koeln_1[7]}). 
The accent color may also be used in lighter shades of 80%, 60%, or 40%.
}
\examples{
uni_koeln_1
seecol(uni_koeln_1, main = "Primary colors of Uni Koeln")  # view palette
demopal(uni_koeln_1, type = 4, 
        main = "Primary colors at the University of Cologne")

}
\seealso{
\code{\link{uni_koeln_2}} for departmental colors of the University of Koeln;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}}
}
\author{
\strong{unikn}, 2022-12-05.
}
\concept{contributed color palettes}
\keyword{datasets}
