% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{uni_koeln_2}
\alias{uni_koeln_2}
\title{Departmental colors of the University of Koeln (Cologne)}
\format{
An object of class \code{character} of length 7.
}
\source{
Color definitions are based on 
\url{https://kommunikation-marketing.uni-koeln.de/marketing/corporate_design/index_ger.html}.
}
\usage{
uni_koeln_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_koeln_2} provides the departmental (faculty) colors 
of the \href{https://www.uni-koeln.de/}{Universitaet zu Koeln}, Germany.
}
\details{
The named colors of \code{uni_koeln_2} correspond to the following departments:

\enumerate{
\item \code{gruen}: Wirtschafts-/Sozialwissenschaftliche Fakultaet
\item \code{bordeaux}: Rechtswissenschaftliche Fakultaet 
\item \code{rot}: Medizinische Fakultaet
\item \code{violett}: Philosophische Fakultaet
\item \code{blau}: Mathematisch-Naturwissenschaftliche Fakultaet
\item \code{orange}: Humanwissenschaftlicheschaftliche Fakultaet 
\item \code{hellblau}: Verwaltung 
}
}
\examples{
uni_koeln_2
seecol(uni_koeln_2, main = "Departmental colors of Uni Koeln")  # view palette
demopal(uni_koeln_2, type = 3, 
        main = "Department colors at the University of Cologne")

}
\seealso{
\code{\link{uni_koeln_1}} for primary colors of the University of Koeln;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}}
}
\author{
\strong{unikn}, 2022-12-05.
}
\concept{contributed color palettes}
\keyword{datasets}
