% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_pals.R
\docType{data}
\name{uni_konstanz}
\alias{uni_konstanz}
\title{Default colors of the University of Konstanz.}
\format{
An object of class \code{character} of length 11.
}
\source{
R package \strong{unikn}, 
retrieved from \url{https://CRAN.R-project.org/package=unikn},    
doi \href{https://doi.org/10.5281/zenodo.7096191}{10.5281/zenodo.7096191}. 

Color definitions are based on 
\url{https://www.uni-konstanz.de/en/university/news-and-media/create-online-and-print-media/corporate-design/the-university-of-konstanzs-corporate-design/}  
\url{https://www.uni-konstanz.de/en/university/news-and-media/create-online-and-print-media/corporate-design/colours-for-complex-graphics/}
}
\usage{
uni_konstanz
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_konstanz} provides the default color palette 
of the \href{https://www.uni-konstanz.de/}{University of Konstanz}, Germany.
}
\details{
The \strong{unikn} package provides many additional color palettes for 
the \href{https://www.uni-konstanz.de/}{University of Konstanz}.
}
\examples{
uni_konstanz
dim(uni_konstanz)     # 1 11
seecol(uni_konstanz)  # view color palette

}
\seealso{
\code{\link{uni_konstanz_pref}} for preferred colors of the University of Konstanz;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{uni_konstanz_pref}}
}
\author{
Hansjoerg Neth, 2022-10-20.
}
\concept{contributed color palettes}
\keyword{datasets}
