% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{uni_princeton_0}
\alias{uni_princeton_0}
\title{Basic colors of Princeton University.}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on 
\url{https://communications.princeton.edu/guides-tools/logo-graphic-identity}.
}
\usage{
uni_princeton_0
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_princeton_0} provides the two basic colors 
of \href{https://www.princeton.edu/}{Princeton University}, NJ, USA.
}
\examples{
uni_princeton_0
seecol(uni_princeton_0)  # view color palette
demopal(uni_princeton_0, type = 1, main = "Colors of Princeton University")

}
\seealso{
\code{\link{uni_princeton_1}} for Princeton University colors (on white backgrounds); 
\code{\link{uni_princeton_2}} for Princeton University colors (on black backgrounds); 
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{eth_pal_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{mpg_pal}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}}
}
\author{
Hansjoerg Neth, 2022-10-20.
}
\concept{contributed color palettes}
\keyword{datasets}
