% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{uni_regensburg_2}
\alias{uni_regensburg_2}
\title{Departmental colors of the University of Regensburg (for areas)}
\format{
An object of class \code{character} of length 12.
}
\source{
Color definitions are based on 
\url{https://www.uni-regensburg.de/verwaltung/corporate-design/index.html}.
}
\usage{
uni_regensburg_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_regensburg_2} provides the departmental (faculty) colors 
for coloring print or web areas 
at the \href{https://www.uni-regensburg.de/}{University of Regensburg}, Germany.
}
\details{
Named colors of \code{uni_regensburg_2} correspond to the following UR faculties:

\enumerate{
\item \code{urangelb}: Rechtswissenschaft
\item \code{laerchennadelgruen}: Wirtschaftswissenschaften
\item \code{vatikangelb}: Katholische Theologie
\item \code{orangerot}: Philosophie, Kunst-, Geschichts- und Gesellschaftswissenschaften
\item \code{heucherarot}: Humanwissenschaften
\item \code{heidenelkenrot}: Sprach-, Literatur- und Kulturwissenschaften
\item \code{blattgruen}: Biologie und Vorklinische Medizin
\item \code{tuerkisgruen}: Mathematik
\item \code{eisvogelblau}: Physik
\item \code{spektralblau}: Chemie und Pharmazie
\item \code{capriblau}: Medizin
\item \code{glutrot}: Informatik und Data Science
}

When coloring large areas, named colors can be used in shades of 70%.
}
\examples{
uni_regensburg_2
seecol(uni_regensburg_2)  # view color palette
demopal(uni_regensburg_2, type = 3, 
        main = "Department colors at the University of Regensburg")

}
\seealso{
\code{\link{uni_regensburg_1}} and \code{\link{uni_regensburg_3}} for alternative colors of the University of Regensburg;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}}
}
\author{
\strong{unikn}, 2022-11-30.
}
\concept{contributed color palettes}
\keyword{datasets}
