% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{uni_regensburg_3}
\alias{uni_regensburg_3}
\title{Departmental colors of the University of Regensburg (for fonts)}
\format{
An object of class \code{character} of length 12.
}
\source{
Color definitions are based on 
\url{https://www.uni-regensburg.de/verwaltung/corporate-design/index.html}.
}
\usage{
uni_regensburg_3
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_regensburg_3} provides the departmental (faculty) colors 
for web fonts 
at the \href{https://www.uni-regensburg.de/}{University of Regensburg}, Germany.
}
\details{
Named colors of \code{uni_regensburg_3} correspond to the following UR faculties:

\enumerate{
\item urangelb: Rechtswissenschaft
\item laerchennadelgruen: Wirtschaftswissenschaften
\item vatikangelb: Katholische Theologie
\item orangerot: Philosophie, Kunst-, Geschichts- und Gesellschaftswissenschaften
\item heucherarot: Humanwissenschaften
\item heidenelkenrot: Sprach-, Literatur- und Kulturwissenschaften
\item blattgruen: Biologie und Vorklinische Medizin
\item tuerkisgruen: Mathematik
\item eisvogelblau: Physik
\item spektralblau: Chemie und Pharmazie
\item capriblau: Medizin
\item glutrot: Informatik und Data Science
}

Note that the font color glutrot is currently identical to the area color of \code{\link{uni_regensburg_2}}.
}
\examples{
uni_regensburg_3
seecol(uni_regensburg_3)  # view color palette
demopal(uni_regensburg_3, type = 4, seed = 13, 
        main = "Department colors at the University of Regensburg")

}
\seealso{
\code{\link{uni_regensburg_1}} and \code{\link{uni_regensburg_2}} for alternative colors of the University of Regensburg;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}}
}
\author{
Hansjoerg Neth, 2022-11-30.
}
\concept{contributed color palettes}
\keyword{datasets}
