% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pal.R
\name{usecol}
\alias{usecol}
\title{Use a color palette.}
\usage{
usecol(pal = pal_unikn)
}
\arguments{
\item{pal}{A color palette (as a vector of colors or color palettes). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{usecol} allows using a color palette \code{pal}.
}
\seealso{
\code{\link{seecol}} to plot color palettes.

Other color palettes: \code{\link{col_scale}},
  \code{\link{mixcol}}, \code{\link{pal_bordeaux}},
  \code{\link{pal_grau}}, \code{\link{pal_karpfenblau}},
  \code{\link{pal_n_sq}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{plot_pal}}, \code{\link{seecol}},
  \code{\link{use_pal_n}}

Other color functions: \code{\link{mixcol}},
  \code{\link{pal_which}}, \code{\link{plot_pal}},
  \code{\link{use_pal_n}}
}
\concept{color functions}
\concept{color palettes}
