% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{xbox}
\alias{xbox}
\title{Plot a box (with x).}
\usage{
xbox(col = unlist(seeblau), dim = c(1, 1))
}
\arguments{
\item{col}{The color to fill the box with (i.e., its background color).
Default: \code{col = unlist(seeblau)}.}

\item{dim}{The x- and y-dimensions of the box. 
Default: \code{dim = c(1, 1)} (i.e., a unit square).}
}
\description{
\code{xbox} plots a box with a cross (x) in its top-right corner.
}
\examples{
xbox()  # default box

}
\seealso{
\code{\link{frame}} to plot a frame.

Other plot functions: \code{\link{frame}}
}
\concept{plot functions}
