% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{ac}
\alias{ac}
\title{Adjust the transparency of a color or color palette.}
\usage{
ac(col, alpha = 0.5, use_names = TRUE)
}
\arguments{
\item{col}{A (required) color or color palette (as a vector).}

\item{alpha}{A factor modifying the opacity alpha 
(as \code{alpha.f} in \code{\link{adjustcolor}}) to a value in \code{[0, 1]}. 
Default: \code{alpha = .50} (i.e., medium opacity).}

\item{use_names}{A logical value indicating whether color names should be adjusted 
to include the values of \code{alpha}. 
Default: \code{use_names = TRUE}.}
}
\value{
A color vector of the same length as \code{col}, 
transformed by \code{link{adjustcolor}}.
}
\description{
\code{ac} adjusts the transparency of a color or color palette \code{col} 
to an opacity level \code{alpha}.
}
\details{
\code{ac} is primarily a convenient wrapper for 
\code{\link{adjustcolor}} of the \strong{grDevices} package, 
but allows for more flexible combinations of 
(multiple) \code{col} and \code{alpha} values.
}
\examples{
 
}
\seealso{
\code{\link{seecol}} to plot color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other color functions: 
\code{\link{grepal}()},
\code{\link{newpal}()},
\code{\link{seecol}()},
\code{\link{shades_of}()},
\code{\link{usecol}()}
}
\concept{color functions}
