% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun_1.R
\name{all_colors}
\alias{all_colors}
\title{A function providing all unikn colors and base R colors.}
\usage{
all_colors(distinct = TRUE)
}
\arguments{
\item{distinct}{Boolean: Return only colors with unique values and names? 
Default: \code{distinct = TRUE} (i.e., remove duplicate values and names).}
}
\value{
A named vector of color values (of type character).
}
\description{
\code{all_colors} combines the \strong{unikn} color gradients with the 
657 named colors provided by \code{\link{colors}} of \strong{grDevices}.
}
\details{
The color values returned (as a named character vector) 
are either HEX/HTML codes or \strong{base} R color names.

The \code{distinct} option removes colors with duplicate color values or names.
}
\keyword{internal}
