% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Bordeaux}
\alias{Bordeaux}
\title{uni.kn color Bordeaux.}
\format{An object of class \code{data.frame} with 1 rows and 1 columns.}
\usage{
Bordeaux
}
\description{
\code{Bordeaux} provides the preferred color of \code{\link{pal_Bordeaux}} 
as an atomic data frame and is defined as 
\code{\link{pal_Bordeaux}[4]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
Bordeaux  # "Bordeaux4" OR "#8E2043"
all.equal(Bordeaux, pal_Bordeaux[4])  # TRUE

}
\seealso{
\code{\link{pal_Bordeaux}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{col_scale}} to extend color palettes (by creating gradients).

Other colors: \code{\link{grau}},
  \code{\link{karpfenblau}}, \code{\link{peach}},
  \code{\link{petrol}}, \code{\link{pinky}},
  \code{\link{seeblau}}, \code{\link{seegruen}}
}
\concept{colors}
\keyword{datasets}
