% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\docType{data}
\name{Bordeaux}
\alias{Bordeaux}
\title{uni.kn color Bordeaux.}
\format{An object of class \code{character} of length 1.}
\usage{
Bordeaux
}
\description{
\code{Bordeaux} provides the preferred color of \code{\link{pal_bordeaux}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_bordeaux}[[4]]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
Bordeaux  # HEX character "#8E2043" (as value)
all.equal(Bordeaux, pal_bordeaux[[4]])  # TRUE (same HEX values)

}
\seealso{
\code{\link{pal_bordeaux}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette.

Other preferred colors: \code{\link{Grau}},
  \code{\link{Karpfenblau}}, \code{\link{Peach}},
  \code{\link{Petrol}}, \code{\link{Pinky}},
  \code{\link{Seeblau}}, \code{\link{Seegruen}},
  \code{\link{Signal}}
}
\concept{preferred colors}
\keyword{datasets}
