% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{col_scale}
\alias{col_scale}
\title{Provide a color gradient scale.}
\usage{
col_scale(col = pal_unikn, ...)
}
\arguments{
\item{col}{A list of colors or color palettes. 
Default: \code{col = \link{pal_unikn}}.}

\item{...}{Additional parameters 
(passed to \code{grDevices::colorRampPalette}).}
}
\description{
\code{col_scale} provides a color gradient 
based on a sequence of input colors or color palettes 
defined by \code{col}.
}
\details{
\code{col_scale} uses \code{grDevices::colorRampPalette} to  
provide a function that takes an integer argument 
(the required number of colors) to  
return a character vector of (rgb) colors 
interpolating color sequence provided by \code{col}.
}
\examples{
# Gradients over 1 color palette:
col_scale()(5)  # 5 colors of default scale pal_unikn
col_scale(pal_petrol)(10)  # 10 colors of pal_petrol

# Gradients over 2 colors:
col_scale(c(seeblau, peach))(10)
col_scale(c(seegruen, petrol))(10)

# Gradients over 2 color palettes:
col_scale(c(pal_seeblau, pal_peach))(10)
col_scale(c(pal_seeblau, pal_karpfenblau))(10)

# Gradients over 3+ colors:
col_scale(c(seeblau, grau, seegruen, peach))(10) 

# Gradients over 3+ color palettes:
col_scale(c(pal_seeblau, pal_grau, pal_petrol))(10)

# Extending exiting color palettes: 
plot_pal(col_scale()(20))
plot_pal(col_scale(pal_seeblau)(10))

# Combining colors to create new palettes:
plot_pal(col_scale(c(pal_seeblau[5], "gold"))(10))
plot_pal(col_scale(c(seeblau, grau, petrol))(10))
plot_pal(col_scale(c(karpfenblau, seeblau, "gold"))(10))
 
# Combining color palettes into new palettes:
plot_pal(col_scale(c(rev(pal_seeblau), pal_petrol))(10)) 
plot_pal(col_scale(c(rev(pal_seeblau), pal_peach))(10)) 

}
\seealso{
\code{\link{plot_pal}} to plot color palettes; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{pal_which}} to get specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_seeblau}},
  \code{\link{pal_seegruen}}, \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
