% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{inst}
\alias{inst}
\title{Primary colors of the University of Koeln (Cologne)}
\format{
An object of class \code{character} of length 36.
}
\usage{
inst
}
\description{
\code{uni_koeln_1} provides the primary colors 
of the \href{https://www.uni-koeln.de/}{Universitaet zu K\u00F6ln}, Germany.
}
\keyword{datasets}
