% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{Karpfenblau}
\alias{Karpfenblau}
\title{uni.kn color Karpfenblau.}
\format{An object of class \code{character} of length 1.}
\usage{
Karpfenblau
}
\description{
\code{Karpfenblau} provides the preferred color of \code{\link{pal_karpfenblau}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_karpfenblau}[[4]]}.
}
\details{
See \url{https://www.uni-konstanz.de/en/university/news-and-media/create-online-and-print-media/corporate-design/} for details.
}
\examples{
Karpfenblau  # HEX character "#3E5496" (as value)
all.equal(Karpfenblau, pal_karpfenblau[[4]])  # TRUE (same HEX values)

seecol(Karpfenblau)  # view color and details

}
\seealso{
\code{\link{pal_karpfenblau}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: \code{\link{Bordeaux}},
  \code{\link{Grau}}, \code{\link{Peach}},
  \code{\link{Petrol}}, \code{\link{Pinky}},
  \code{\link{Seeblau}}, \code{\link{Seegruen}},
  \code{\link{Signal}}
}
\concept{preferred colors}
\keyword{datasets}
