% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text_calls.R
\name{mark}
\alias{mark}
\title{\code{mark} plots 1 or more text strings (provided as a character vector \code{lbls}) 
to an (existing or new) plot and places a colored box behind
each label to mark or highlight it (i.e., make it stand out from the background).}
\usage{
mark(lbls, x = 0, y = 0.55, y_layout = "even", col = "black",
  col_bg = Seeblau, cex = 2, font = 2, new_plot = "none")
}
\arguments{
\item{lbls}{A character vector specifying the text labels 
to be written.}

\item{x}{A numeric vector of x-coordinates at which the 
text labels in \code{lbls} should be written. 
If the lengths of \code{x} and \code{y} differ, 
the shorter one is recycled.}

\item{y}{A numeric vector of y-coordinates at which the 
text labels in \code{lbls} should be written. 
If the lengths of \code{x} and \code{y} differ, 
the shorter one is recycled.}

\item{y_layout}{A numeric value or vector for the vertical 
spacing of labels in \code{lbls}. 
2 special values are \code{"even"} and \code{"flush"} 
(i.e., \code{y_layout = 0}).}

\item{col}{The color(s) of the text label(s). 
Default: \code{col_lbl = "black"}.}

\item{col_bg}{The color(s) to highlight or fill the rectangle(s) with. 
Default: \code{col_bg = Seeblau}.}

\item{cex}{A numeric character expansion factor, 
multiplied by \code{par("cex")} to yield the character size. 
Default: \code{cex = 2}.}

\item{font}{The font to be used. 
Default: \code{font = 2} (i.e., bold).}

\item{new_plot}{Should a new plot be generated? 
Set to \code{"blank"} or \code{"slide"} to create a new plot. 
Default: \code{new_plot = "none"} (i.e., add to existing plot).}
}
\description{
Text formatting parameters (like \code{col}, \code{col_bg}, \code{cex}, \code{font})         
are recycled to match \code{length(lbls)}.
}
\details{
\code{mark} uses the base graphics system \code{graphics::}.
}
\examples{
mark(lbls = "Test text", new_plot = "blank")
                        
}
\seealso{
\code{\link{slide}} and \code{\link{xbox}} to create simple plots (without text).

Other text functions: \code{\link{box_text}},
  \code{\link{plot_box_txt}}
}
\concept{text functions}
