% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{pal_n}
\alias{pal_n}
\title{Get n dedicated colors of a known color palette.}
\usage{
pal_n(n = "all", pal = pal_unikn)
}
\arguments{
\item{n}{A number specifying the desired number colors of pal (as a number) 
or the character string \code{"all"} (to get all colors of \code{pal}). 
Default: \code{n = "all"}.}

\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{pal_n} returns \code{n} dedicated colors of a known color palette \code{pal}.
}
\examples{
pal_n()                   # all colors of default color palette
pal_n(n = 4)              # 4 colors of default color palette
pal_n(pal = pal_seeblau)  # all colors of another color palette

# Default pal:
pal_n(1)   # n = 1 color of default color pal
pal_n(2)
pal_n(4)
pal_n(6)
pal_n(8)
pal_n(10)  # n = 10 colors of default color pal

pal_n(99)  # returns maximal number of colors: pal_unikn_plus (11)

## Other color palettes:
# pal_n(3,  RColorBrewer::brewer.pal(n = 5, name = "Blues"))
# pal_n(10, RColorBrewer::brewer.pal(n = 5, name = "Blues"))

}
\seealso{
\code{\link{pal_which}} for specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_peach}}, \code{\link{pal_petrol}},
  \code{\link{pal_pinky}}, \code{\link{pal_seeblau}},
  \code{\link{pal_seegruen}}, \code{\link{pal_signal}},
  \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{plot_pal}}, \code{\link{seecol}}

Other color functions: \code{\link{pal_which}},
  \code{\link{plot_pal}}
}
\concept{color functions}
\concept{color palettes}
