% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{pal_n_sq}
\alias{pal_n_sq}
\title{Get n^2 dedicated colors of a color palette.}
\usage{
pal_n_sq(n = "all", pal = pal_unikn)
}
\arguments{
\item{n}{A number specifying the desired number colors of pal (as a number) 
or the character string \code{"all"} (to get all colors of \code{pal}). 
Default: \code{n = "all"}.}

\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{pal_n_sq} returns \code{n^2} dedicated colors of a color palette \code{pal} 
(up to a maximum of \code{n + 1} colors).
}
\examples{
pal_n_sq(1)
pal_n_sq(2)
pal_n_sq(3) #  9 colors
pal_n_sq(4) # 11 colors = maximum of (n + 1) colors.

}
\seealso{
\code{\link{pal_n}} for \code{n} dedicated colors of a known color palette; 
\code{\link{pal_which}} for specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{pal_n}},
  \code{\link{pal_peach}}, \code{\link{pal_seeblau}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
