% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pal_seeblau}
\alias{pal_seeblau}
\title{uni.kn seeblau color palette.}
\format{An object of class \code{data.frame} with 1 rows and 5 columns.}
\usage{
pal_seeblau
}
\description{
\code{pal_seeblau} provides an additional uni.kn color palette  
as a data frame containing 5 colors.
}
\details{
See https://www.uni-konstanz.de for details.
}
\examples{
pal_seeblau
dim(pal_seeblau)  # 1 5
pal_seeblau[3]    # preferred (named) color
pal_seeblau[[3]]  # preferred color "seeblau3" = "#59C7EB"

}
\seealso{
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_n}} for \code{n} dedicated colors of a known color palette.

Other color palettes: \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
