% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pal_unikn}
\alias{pal_unikn}
\title{uni.kn default color palette.}
\format{An object of class \code{data.frame} with 1 rows and 10 columns.}
\usage{
pal_unikn
}
\description{
\code{pal_unikn} provides the default uni.kn color palette  
as a data frame containing 10 colors.
}
\details{
This is the primary (web/sRGB) scale.

See https://www.uni-konstanz.de for details.
}
\examples{
pal_unikn
dim(pal_unikn)  # 1 10
pal_unikn[1]    # named color
pal_unikn[[1]]  # color

}
\seealso{
\code{\link{pal_unikn_ppt}} for an alternative (ppt) version; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_n}} for \code{n} dedicated colors of a known color palette.

Other color palettes: \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_seeblau}}, \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_which}},
  \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
