% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def.R
\docType{data}
\name{pal_unikn}
\alias{pal_unikn}
\title{uni.kn default color palette.}
\format{An object of class \code{data.frame} with 1 rows and 10 columns.}
\usage{
pal_unikn
}
\description{
\code{pal_unikn} provides the default uni.kn color palette  
as a data frame containing 10 colors.
}
\details{
This is the primary (web/sRGB) scale.

See \url{https://www.uni-konstanz.de} for details.
}
\examples{
pal_unikn
dim(pal_unikn)  # 1 10

# Access by position:
pal_unikn[2]    # 2nd named color "seeblau3" (as df)
pal_unikn[[2]]  # 2nd color value "#59C7EB"

# Access by name: 
pal_unikn["seeblau3"]    # color "seeblau3" (as df)
pal_unikn[["seeblau3"]]  # color value "#59C7EB"

# Plotting palette:
seepal(pal_unikn)

}
\seealso{
\code{\link{pal_unikn_ppt}} for an alternative (ppt) version; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{seepal}} to show and use color palettes.

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_which}},
  \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
