% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\docType{data}
\name{pal_unikn_dark}
\alias{pal_unikn_dark}
\title{uni.kn dark colors in a color palette.}
\format{An object of class \code{data.frame} with 1 rows and 8 columns.}
\usage{
pal_unikn_dark
}
\description{
\code{pal_unikn_dark} provides an additional uni.kn color palette  
that collects 2 dark colors of 4 color palettes  
as a data frame containing 8 colors (in 4 pairs).
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
pal_unikn_dark
dim(pal_unikn_dark)  # 1 8 
pal_unikn_dark[1]    # color "karpfenblau5" by position
pal_unikn_dark[[1]]  # color value by position: #324376"
pal_unikn_dark["karpfenblau5"]  # color value by name

# Plotting palette:
plot_pal(pal_unikn_dark)

}
\seealso{
\code{\link{pal_unikn_light}} for a lighter uni.kn color palette; 
\code{\link{pal_unikn_pair}} for a pairwise uni.kn color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{col_scale}} to extend color palettes (by creating gradients).

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_Bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
