% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pal_unikn_ppt}
\alias{pal_unikn_ppt}
\title{uni.kn color scale (ppt version).}
\format{An object of class \code{data.frame} with 1 rows and 10 columns.}
\usage{
pal_unikn_ppt
}
\description{
\code{pal_unikn_ppt} provides an uni.kn color scale 
as a data frame containing 10 colors.
}
\details{
This is a secondary (ppt) variant with more muted colors.

See \code{\link{pal_unikn}} for the primary/default (web/sRGB) scale 
and https://www.uni-konstanz.de for details.
}
\examples{
pal_unikn_ppt
dim(pal_unikn_ppt)  # 1 10
pal_unikn_ppt[1]    # named color
pal_unikn_ppt[[1]]  # color

}
\seealso{
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color scales: \code{\link{col_n_sq}},
  \code{\link{col_n}}, \code{\link{col_which}},
  \code{\link{pal_unikn}}, \code{\link{plot_pal}}
}
\concept{color scales}
\keyword{datasets}
