% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\docType{data}
\name{pal_unikn_pref}
\alias{pal_unikn_pref}
\title{uni.kn preferred colors in a color palette.}
\format{An object of class \code{data.frame} with 1 rows and 9 columns.}
\usage{
pal_unikn_pref
}
\description{
\code{pal_unikn_pref} provides an additional uni.kn color palette  
that collects the preferred color of each palette 
as a data frame containing 9 (or 8 + 1) colors.
}
\details{
Note that the color \code{\link{signal}} is not identified 
as a preferred color according to the official color definition.

See \url{https://www.uni-konstanz.de} for details.
}
\examples{
pal_unikn_pref
dim(pal_unikn_pref)  # 1 X (2)
pal_unikn_pref[1]    # preferred (named) color
pal_unikn_pref[[1]]  # preferred color value: #59C7EB"
pal_unikn_pref["seeblau"]  # preferred color by name

# Plotting palette:
plot_pal(pal_unikn_pref)

}
\seealso{
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{col_scale}} to extend color palettes (by creating gradients).

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_Bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}, \code{\link{plot_pal}}
}
\concept{color palettes}
\keyword{datasets}
