% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{pal_which}
\alias{pal_which}
\title{Get specific colors of a color palette.}
\usage{
pal_which(which = "all", pal = pal_unikn)
}
\arguments{
\item{which}{A numeric vector specifying the desired colors of pal (by position) 
or the character string \code{"all"} (to get all colors of \code{pal}). 
Default: \code{which = "all"}.}

\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{pal_which} returns the colors of a color palette \code{pal} 
that are specified in a numeric vector \code{which}.
}
\examples{
pal_which()  # returns default color palette
pal_which(which = "all")
pal_which(which = 1:3)
pal_which(which = rep(1, 20))

}
\seealso{
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{col_scale}} to extend color palettes (by creating gradients).

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_Bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{plot_pal}}
}
\concept{color palettes}
