% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\docType{data}
\name{petrol}
\alias{petrol}
\title{uni.kn color petrol.}
\format{An object of class \code{data.frame} with 1 rows and 1 columns.}
\usage{
petrol
}
\description{
\code{petrol} provides the preferred color of \code{\link{pal_petrol}} 
as an atomic data frame and is defined as 
\code{\link{pal_petrol}[4]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
petrol  # "petrol" OR "#077187"
all.equal(petrol, pal_petrol[4])  # TRUE

}
\seealso{
\code{\link{pal_petrol}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette.

Other colors: \code{\link{Bordeaux}},
  \code{\link{Seeblau}}, \code{\link{grau}},
  \code{\link{karpfenblau}}, \code{\link{peach}},
  \code{\link{pinky}}, \code{\link{seeblau}},
  \code{\link{seegruen}}, \code{\link{signal}}
}
\concept{colors}
\keyword{datasets}
