% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pinky}
\alias{pinky}
\title{uni.kn color pinky.}
\format{An object of class \code{data.frame} with 1 rows and 1 columns.}
\usage{
pinky
}
\description{
\code{pinky} provides the preferred color of \code{\link{pal_pinky}} 
as an atomic data frame and is defined as 
\code{\link{pal_pinky}[4]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
pinky  # "pinky4" OR "#E0607E"
all.equal(pinky, pal_pinky[4])  # TRUE

}
\seealso{
\code{\link{pal_pinky}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{col_scale}} to extend color palettes (by creating gradients).

Other colors: \code{\link{Bordeaux}}, \code{\link{grau}},
  \code{\link{karpfenblau}}, \code{\link{peach}},
  \code{\link{petrol}}, \code{\link{seeblau}},
  \code{\link{seegruen}}
}
\concept{colors}
\keyword{datasets}
