% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\docType{data}
\name{Pinky}
\alias{Pinky}
\title{uni.kn color Pinky.}
\format{An object of class \code{character} of length 1.}
\usage{
Pinky
}
\description{
\code{Pinky} provides the preferred color of \code{\link{pal_pinky}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_pinky}[[4]]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
Pinky  # HEX character "#E0607E" (as value)
all.equal(Pinky, pal_pinky[[4]])  # TRUE (same HEX values)

}
\seealso{
\code{\link{pal_pinky}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette.

Other colors: \code{\link{Grau}},
  \code{\link{Karpfenblau}}, \code{\link{Peach}},
  \code{\link{Petrol}}, \code{\link{Seeblau}},
  \code{\link{Seegruen}}, \code{\link{bordeaux}},
  \code{\link{grau}}, \code{\link{karpfenblau}},
  \code{\link{peach}}, \code{\link{petrol}},
  \code{\link{pinky}}, \code{\link{seeblau}},
  \code{\link{seegruen}}, \code{\link{signal}}

Other preferred colors: \code{\link{Grau}},
  \code{\link{Karpfenblau}}, \code{\link{Peach}},
  \code{\link{Petrol}}, \code{\link{Seeblau}},
  \code{\link{Seegruen}}, \code{\link{seeblau}}
}
\concept{colors}
\concept{preferred colors}
\keyword{datasets}
