% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R, R/plot_pal.R
\name{plot_pal}
\alias{plot_pal}
\title{Plot a color palette.}
\usage{
plot_pal(pal = pal_unikn)

plot_pal(pal = pal_unikn)
}
\arguments{
\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}

\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}
}
\description{
\code{plot_pal} plots information on a color palette \code{pal}.

\code{plot_pal} plots information on a color palette \code{pal}.
}
\examples{
plot_pal()  # plots default color palette 
plot_pal(pal_unikn_plus) 

## Not run: 

## Plotting palettes from other packages: ------ 

## RColorBrewer palettes: 
# plot_pal(RColorBrewer::brewer.pal(n = "all", name = "Set1"))
# plot_pal(RColorBrewer::brewer.pal(n = 5, name = "Blues"))

## viridisLite palettes: 
# plot_pal(viridisLite::viridis(20))
# plot_pal(viridisLite::inferno(25)) 
# plot_pal(viridisLite::magma(25))
# plot_pal(viridisLite::plasma(25))
  
## yarrr palettes: 
# plot_pal(yarrr::piratepal(palette = "appletv"))
# plot_pal(yarrr::piratepal(palette = "google"))

## End(Not run)

plot_pal()  # plots default color palette 
plot_pal(pal_unikn_plus) 

## Not run: 

# Plotting palettes from other packages: ------ 

## RColorBrewer palettes: 
# plot_pal(RColorBrewer::brewer.pal(n = "all", name = "Set1"))
# plot_pal(RColorBrewer::brewer.pal(n = 5, name = "Blues"))

## viridisLite palettes: 
# plot_pal(viridisLite::viridis(20))
# plot_pal(viridisLite::inferno(25)) 
# plot_pal(viridisLite::magma(25))
# plot_pal(viridisLite::plasma(25))
  
## yarrr palettes: 
# plot_pal(yarrr::piratepal(palette = "appletv"))
# plot_pal(yarrr::piratepal(palette = "google"))

## End(Not run)

}
\seealso{
\code{\link{col_scale}} to extend color palettes (by creating gradients);  
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{pal_which}} to get specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

\code{\link{col_scale}} to extend color palettes (by creating gradients);  
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{pal_which}} to get specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{pal_which}}
}
\concept{color palettes}
