% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_definitions.R
\docType{data}
\name{Seeblau}
\alias{Seeblau}
\title{uni.kn color Seeblau.}
\format{An object of class \code{character} of length 1.}
\usage{
Seeblau
}
\description{
\code{Seeblau} provides the preferred color of \code{\link{pal_seeblau}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_seeblau}[[3]]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
Seeblau  # HEX character "#59C7EB" (as value)
all.equal(Seeblau, pal_seeblau[[3]])  # TRUE (both values)

}
\seealso{
\code{\link{pal_seeblau}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette.

Other colors: \code{\link{Bordeaux}}, \code{\link{grau}},
  \code{\link{karpfenblau}}, \code{\link{peach}},
  \code{\link{petrol}}, \code{\link{pinky}},
  \code{\link{seeblau}}, \code{\link{seegruen}},
  \code{\link{signal}}

Other preferred colors: \code{\link{seeblau}}
}
\concept{colors}
\concept{preferred colors}
\keyword{datasets}
