% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{Seeblau}
\alias{Seeblau}
\title{uni.kn color Seeblau.}
\format{An object of class \code{character} of length 1.}
\usage{
Seeblau
}
\description{
\code{Seeblau} provides the preferred color of \code{\link{pal_seeblau}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_seeblau}[[3]]}.
}
\details{
See \url{https://www.uni-konstanz.de/en/university/news-and-media/create-online-and-print-media/corporate-design/} for details.
}
\examples{
Seeblau  # HEX character "#59C7EB" (as value)
all.equal(Seeblau, pal_seeblau[[3]])  # TRUE (same HEX values)

seecol(Seeblau)  # view color and details

}
\seealso{
\code{\link{pal_seeblau}} for the corresponding color palette; 
\code{\link{pal_unikn}} for the unikn default color palette with all 5 colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{Bordeaux}},
\code{\link{Grau}},
\code{\link{Karpfenblau}},
\code{\link{Peach}},
\code{\link{Petrol}},
\code{\link{Pinky}},
\code{\link{Seegruen}},
\code{\link{Signal}}
}
\concept{preferred colors}
\keyword{datasets}
