% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pal.R
\name{seecol}
\alias{seecol}
\alias{seepal}
\title{Show and use color palettes.}
\usage{
seecol(pal = "all", n = "all", hex = NULL, rgb = NULL,
  col_brd = NULL, grid = TRUE, ...)
}
\arguments{
\item{pal}{A color palette (as a list of colors) or 
the character string \code{"all"} (for showing all 
color palettes in the current environment). 
Default: \code{pal = "all"}.}

\item{n}{Number of colors to show or use. 
If \code{n} is lower or higher than the length of the current 
color palette \code{pal}, the color palette is reduced or extrapolated 
(using \code{grDevices::colorRampPalette}). 
Default: \code{n = "all"}.}

\item{hex}{Should HEX color values be shown?
Default: \code{hex = NULL} (i.e., show HEX color values 
when there is sufficient space to print them).}

\item{rgb}{Should RGB color values be shown?
Default: \code{rgb = NULL} (i.e., show RGB color values 
when there is sufficient space to print them).}

\item{col_brd}{Color of box borders (if shown). 
Default: \code{col_brd = NULL}.}

\item{grid}{Show grid?
Default: \code{grid = TRUE}.}

\item{...}{Other graphical parameters 
(passed to \code{plot_col}).}
}
\description{
\code{seecol} provides an interface to seeing colors 
or color palettes.
}
\details{
\code{seecol} has 2 main modes, based on the contents of its \code{pal} argument:

\enumerate{

\item \code{pal} is "all" (or another list of multiple color palettes):

Plot vectors of all current color palettes for comparing them.

\item \code{pal} is one specific color palette (or a vector of multiple colors or color palettes):

Plot the current color palette and optional details on its colors.

}
}
\examples{
# See all color palettes: 
seecol()  # same as seecol(pal = "all") 

# See details of a color palette: 
seecol(pal_unikn_plus)  # see a specific color palette

# Combining colors or color palettes: 
seecol(c("black", "firebrick", "gold"))
seecol(c(rev(pal_seeblau), pal_seegruen))
seecol(c(rev(pal_seeblau), "white", pal_pinky))

# Using n to reduce or extend color palettes:
seecol(n =  3)  
seecol(n = 12)

seecol(pal_unikn_plus, n = 5)
seecol(pal_seeblau, n = 10)

# Combining and extending color palettes: 
seecol(c(pal_seeblau, "white", pal_bordeaux), n = 21)

}
\seealso{
\code{\link{col_scale}} to extend color palettes (by creating gradients);  
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{pal_which}} to get specific colors of a color palette; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{col_scale}},
  \code{\link{pal_bordeaux}}, \code{\link{pal_grau}},
  \code{\link{pal_karpfenblau}}, \code{\link{pal_n_sq}},
  \code{\link{pal_n}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{plot_pal}}
}
\concept{color palettes}
