% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_1.R
\docType{data}
\name{Seegruen}
\alias{Seegruen}
\title{uni.kn color Seegruen.}
\format{An object of class \code{character} of length 1.}
\usage{
Seegruen
}
\description{
\code{Seegruen} provides the preferred color of \code{\link{pal_seegruen}} 
(as an atomic HEX character value) and is defined as 
\code{\link{pal_seegruen}[[4]]}.
}
\details{
See \url{https://www.uni-konstanz.de/en/university/news-and-media/create-online-and-print-media/corporate-design/} for details.
}
\examples{
Seegruen  # HEX character "#0A9086" (as value)
all.equal(Seegruen, pal_seegruen[[4]])  # TRUE (same HEX values)

seecol(Seegruen)  # view color and details

}
\seealso{
\code{\link{pal_seegruen}} for the corresponding color palette; 
\code{\link{pal_unikn}} for the unikn default color palette with all 5 colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{Bordeaux}},
\code{\link{Grau}},
\code{\link{Karpfenblau}},
\code{\link{Peach}},
\code{\link{Petrol}},
\code{\link{Pinky}},
\code{\link{Seeblau}},
\code{\link{Signal}}
}
\concept{preferred colors}
\keyword{datasets}
