% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{seegruen}
\alias{seegruen}
\title{uni.kn color seegruen.}
\format{An object of class \code{data.frame} with 1 rows and 1 columns.}
\usage{
seegruen
}
\description{
\code{seegruen} provides the preferred color of \code{\link{pal_seegruen}} 
as an atomic data frame and is defined as 
\code{\link{pal_seegruen}[4]}.
}
\details{
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
seegruen  # "seegruen4" OR "#0A9086"
all.equal(seegruen, pal_seegruen[4])  # TRUE

}
\seealso{
\code{\link{pal_seegruen}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette.

Other colors: \code{\link{Bordeaux}}, \code{\link{grau}},
  \code{\link{karpfenblau}}, \code{\link{peach}},
  \code{\link{petrol}}, \code{\link{pinky}},
  \code{\link{seeblau}}, \code{\link{signal}}
}
\concept{colors}
\keyword{datasets}
