% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{signal}
\alias{signal}
\title{uni.kn color signal.}
\format{An object of class \code{data.frame} with 1 rows and 1 columns.}
\usage{
signal
}
\description{
\code{signal} provides the alert color of \code{\link{pal_signal}} 
as an atomic data frame and is defined as 
\code{\link{pal_signal}[2]}.
}
\details{
The official specification of \code{\link{pal_signal}} 
does not identify a preferred color. 
We provide \code{\link{pal_signal}[2]} as a dedicated color 
as it is well-suited for creating color gradients 
(see \code{\link{col_scale}}). 
 
See \url{https://www.uni-konstanz.de} for details.
}
\examples{
signal  # "signal2" OR "#EFDC60"
all.equal(signal, pal_signal[2])  # TRUE

}
\seealso{
\code{\link{pal_signal}} for the corresponding color palette; 
\code{\link{pal_unikn_pref}} for a uni.kn color palette with all preferred colors; 
\code{\link{pal_unikn}} for the default uni.kn color palette; 
\code{\link{pal_unikn_plus}} for a uni.kn color palette with all colors of \code{\link{pal_seeblau}}; 
\code{\link{pal_n}} to get \code{n} dedicated colors of a known color palette; 
\code{\link{col_scale}} to extend color palettes (by creating gradients).

Other colors: \code{\link{Bordeaux}}, \code{\link{grau}},
  \code{\link{karpfenblau}}, \code{\link{peach}},
  \code{\link{petrol}}, \code{\link{pinky}},
  \code{\link{seeblau}}, \code{\link{seegruen}}
}
\concept{colors}
\keyword{datasets}
