% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_unikn}
\alias{theme_unikn}
\title{Basic unikn theme (for ggplot2)}
\usage{
theme_unikn(
  col_title = pal_seeblau[[4]],
  col_line = grey(0.05, 1),
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  axis_ticks_mm = 2
)
}
\arguments{
\item{col_title}{Color of plot title (optional).
Default: \code{col_title = pal_seeblau[[4]]}. 
Consider using \code{col_title = "black"} when data uses 
\code{Seeblau} colors.}

\item{col_line}{Color of line elements 
used for axis lines and ticks, strip and panel border.
Default: \code{col_line = grey(.05, 1)} (i.e., dark grey).}

\item{base_size}{Base font size (optional, numeric). 
Default: \code{base_size = 11}.}

\item{base_family}{Base font family (optional, character). 
Default: \code{base_family = ""}. 
Options include \code{"mono"}, \code{"sans"} (default), and "serif".}

\item{base_line_size}{Base line size (optional, numeric). 
Default: \code{base_line_size = base_size/22}.}

\item{base_rect_size}{Base rectangle size (optional, numeric). 
Default: \code{base_rect_size = base_size/22}.}

\item{axis_ticks_mm}{Length of axis ticks (in mm). 
Default: \code{axis_ticks_mm = 2}.}
}
\description{
\code{theme_unikn} provides a basic \bold{unikn} theme 
to use in \bold{ggplot2} commands.
}
\details{
The theme is lightweight and no-nonsense, but somewhat 
opinionated (e.g., in using mostly grey scales to 
allow emphasizing data points with color accents).
}
\examples{

\donttest{
  # Plotting iris dataset (using ggplot2, theme_unikn, and unikn colors):
  
  library('ggplot2')  # theme_unikn requires ggplot2 
  
  ggplot(datasets::iris) +
    geom_jitter(aes(x = Petal.Length, y = Petal.Width, color = Species), size = 3, alpha = 2/3) +
    facet_wrap(~Species) +
    scale_color_manual(values = usecol(pal = c(Pinky, Seeblau, Seegruen))) +
    labs(tag = "A", title = "Iris petals by species (using theme_unikn)",
         caption = "Data from datasets::iris") + 
    theme_unikn()
}

}
\seealso{
\code{\link{theme_grau}} for a grey theme; 
\code{\link{theme_bwkn}} for a bw theme; 
\code{\link{theme_minikn}} for a minimal theme.

Other plot functions: 
\code{\link{slide}()},
\code{\link{theme_bwkn}()},
\code{\link{theme_grau}()},
\code{\link{theme_minikn}()},
\code{\link{xbox}()}
}
\concept{plot functions}
