% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{uni_konstanz_pref}
\alias{uni_konstanz_pref}
\title{Preferred colors of the University of Konstanz}
\format{
An object of class \code{character} of length 9.
}
\source{
R package \strong{unikn}, 
retrieved from \url{https://CRAN.R-project.org/package=unikn},    
doi \doi{10.5281/zenodo.7096191}. 

Color definitions are based on 
\url{https://www.uni-konstanz.de/en/university/news-and-media/create-online-and-print-media/corporate-design/the-university-of-konstanzs-corporate-design/}  
\url{https://www.uni-konstanz.de/en/university/news-and-media/create-online-and-print-media/corporate-design/colours-for-complex-graphics/}
}
\usage{
uni_konstanz_pref
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_konstanz_pref} provides the color palette of preferred colors  
of the \href{https://www.uni-konstanz.de/}{University of Konstanz}, Germany.
}
\details{
The \strong{unikn} package provides many additional color palettes for 
the \href{https://www.uni-konstanz.de/}{University of Konstanz}.
}
\examples{
uni_konstanz_pref
seecol(uni_konstanz_pref)  # view color palette
demopal(uni_konstanz_pref, type = 3, main = "Preferred colors of the Uni Konstanz")

}
\seealso{
\code{\link{uni_konstanz}} for the default colors of the University of Konstanz;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{eth_pal_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{mpg_pal}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_konstanz}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}}
}
\author{
Hansjoerg Neth, 2022-10-20.
}
\concept{contributed color palettes}
\keyword{datasets}
