% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pals.R
\docType{data}
\name{uni_regensburg_1}
\alias{uni_regensburg_1}
\title{General colors of the University of Regensburg.}
\format{
An object of class \code{character} of length 12.
}
\source{
Color definitions are based on 
\url{https://www.uni-regensburg.de/verwaltung/corporate-design/index.html}.
}
\usage{
uni_regensburg_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_regensburg_1} provides the general (administrative and service-oriented) colors 
of the \href{https://www.uni-regensburg.de/}{University of Regensburg}, Germany.
}
\details{
Named colors of \code{uni_regensburg_1} correspond to the following services:

\enumerate{
\item neutralgrau: UR-Hausfarbe, Logo 
\item blaugrau: Leitung, Organe und Verwaltung (LOV, Ausnahme: Referate I/3, I/4, I/5, I/6)
\item schwarzrot: Chancengleichheit und Familie (Familien-Service, Frauenbeauftragte, Frauenfoerderung, Gleichstellung)
\item schilfgruen: Service-Einrichtungen der Verwaltung für Studierende (Akademisches Auslandsamt, Studentenkanzlei, Zentrales Pruefungssekretariat, Zentrale Studienberatung, Senatsbeauftragter für Behinderte und chronisch Kranke) 
\item dianthusrot: Universitaetsbibliothek (UB)
\item saphirblau: Rechenzentrum (RZ)
\item farngruen: Zentrum fuer Weiterbildung
\item rotbraun: Zentrum fuer Sprache und Kommunikation (ZSK)
\item perlviolett: Europaeum (Ost-West-Zentrum)
\item violettrot: Zentrum fuer Hochschul- und Wissenschaftsdidaktik (ZHW)
\item gruenbraun: Regensburger Universitätszentrum fuer Lehrerbildung (RUL)
\item wasserblau: Sportzentrum (SZ)
}

When coloring large areas, named colors can be used in shades of 70%.
}
\examples{
uni_regensburg_1
seecol(uni_regensburg_1)  # view color palette
demopal(uni_regensburg_1, alpha = .70, type = 2, 
        main = "General colors of the University of Regensburg")

}
\seealso{
\code{\link{uni_regensburg_2}} and \code{\link{uni_regensburg_3}} for the faculty colors of the University of Regensburg;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other contributed color palettes: 
\code{\link{eth_pal_grey}},
\code{\link{eth_pal_light}},
\code{\link{eth_pal}},
\code{\link{mpg_pal}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_konstanz_pref}},
\code{\link{uni_konstanz}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}}
}
\author{
Hansjoerg Neth, 2022-11-25.
}
\concept{contributed color palettes}
\keyword{datasets}
