% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{url_unikn}
\alias{url_unikn}
\title{\code{url_unikn} formats an URL the uni.kn way.}
\usage{
url_unikn(url = "https://www.uni-konstanz.de/")
}
\arguments{
\item{url}{The url to be written (as copied from a web browser).}
}
\description{
\code{url_unikn} removes various patterns (e.g., 
\code{"http", "https", "://", "www."}) from the 
front of a given URL and returns the remaining character string 
with a figure dash prefix.
}
\examples{
url_unikn("https://www.uni-konstanz.de/")

}
\seealso{
\code{\link{xbox}} to create a new xbox (without text).

Other text functions: \code{\link{mark}},
  \code{\link{post}}, \code{\link{uline}}
}
\concept{text functions}
