% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{use_pal_n}
\alias{use_pal_n}
\title{Get n dedicated colors of a known color palette.}
\usage{
use_pal_n(pal = pal_unikn, n = "all")
}
\arguments{
\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_unikn}}.}

\item{n}{A number specifying the desired number colors of pal (as a number) 
or the character string \code{"all"} (to get all colors of \code{pal}). 
Default: \code{n = "all"}.}
}
\description{
\code{use_pal_n} returns \code{n} dedicated colors of a known color palette \code{pal}.
}
\examples{
# use_pal_n()                   # all colors of default color palette
# use_pal_n(n = 4)              # 4 colors of default color palette
# use_pal_n(pal = pal_seeblau)  # all colors of another color palette

## Default pal:
# use_pal_n(1)   # n = 1 color of default color pal
# use_pal_n(2)
# use_pal_n(4)
# use_pal_n(6)
# use_pal_n(8)
# use_pal_n(10)  # n = 10 colors of default color pal

# use_pal_n(99)  # returns maximal number of colors: pal_unikn_plus (11)

## Other color palettes:
# use_pal_n(3,  RColorBrewer::brewer.pal(n = 5, name = "Blues"))
# use_pal_n(10, RColorBrewer::brewer.pal(n = 5, name = "Blues"))

}
\seealso{
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color palettes: \code{\link{col_scale}},
  \code{\link{mixcol}}, \code{\link{pal_bordeaux}},
  \code{\link{pal_grau}}, \code{\link{pal_karpfenblau}},
  \code{\link{pal_n_sq}}, \code{\link{pal_peach}},
  \code{\link{pal_petrol}}, \code{\link{pal_pinky}},
  \code{\link{pal_seeblau}}, \code{\link{pal_seegruen}},
  \code{\link{pal_signal}}, \code{\link{pal_unikn_dark}},
  \code{\link{pal_unikn_light}},
  \code{\link{pal_unikn_pair}},
  \code{\link{pal_unikn_plus}},
  \code{\link{pal_unikn_ppt}},
  \code{\link{pal_unikn_pref}}, \code{\link{pal_unikn}},
  \code{\link{plot_pal}}, \code{\link{seecol}},
  \code{\link{usecol}}

Other color functions: \code{\link{mixcol}},
  \code{\link{pal_which}}, \code{\link{plot_pal}},
  \code{\link{usecol}}
}
\concept{color functions}
\concept{color palettes}
