% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{usecol}
\alias{usecol}
\title{Use a color palette.}
\usage{
usecol(pal = pal_unikn, n = "all", use_col_ramp = FALSE)
}
\arguments{
\item{pal}{A color palette (as a vector of colors or color palettes).
Default: \code{pal = \link{pal_unikn}}.}

\item{n}{An integer value specifying the desired number of colors from the palette.
For most palettes defined within unikn it uses a predefined selection of 
colors if the desired number of colors is smaller than the available number.  
For all other palettes and \code{n} larger than \code{length(pal)} it uses
\code{\link{colorRampPalette}}.}

\item{use_col_ramp}{A logical value specifying, whether the default of using preselected colors
should be overridden and \code{\link{colorRampPalette}} should always be used to process \code{n}.}
}
\description{
\code{usecol} allows using a color palette \code{pal}.
}
\seealso{
\code{\link{seecol}} to plot color palettes; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color functions: \code{\link{col_scale}},
  \code{\link{mixcol}}, \code{\link{pal_which}},
  \code{\link{plot_pal}}, \code{\link{seecol}},
  \code{\link{use_pal_n}}
}
\concept{color functions}
