% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\name{usecol}
\alias{usecol}
\title{Use a color palette.}
\usage{
usecol(pal = pal_unikn, n = "all", use_names = FALSE,
  use_col_ramp = FALSE)
}
\arguments{
\item{pal}{A color palette (as a vector of colors or color palettes).
Default: \code{pal = \link{pal_unikn}}.}

\item{n}{An integer value specifying the desired number of colors from the palette.
For all palettes defined within unikn by default it uses a pre-defined selection of 
colors if the desired number of colors is smaller than the available number.  
For all other palettes and \code{n} larger than \code{length(pal)} it exteds the palette using
\code{\link{colorRampPalette}}.}

\item{use_names}{A logical value indicating, whether colors should be returned as a named vector.
(defaults to \code{FALSE} for compatibility with ggplot)}

\item{use_col_ramp}{A logical value specifying, whether the default of using pre-selected colors
should be overridden and \code{\link{colorRampPalette}} should always be used to process \code{n}.}
}
\description{
\code{usecol} allows using a color palette \code{pal} (e.g., for plotting).
}
\seealso{
\code{\link{seecol}} to plot color palettes; 
\code{\link{pal_unikn}} for the default uni.kn color palette.

Other color functions: \code{\link{seecol}}
}
\concept{color functions}
