% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box_calls.R
\name{xbox}
\alias{xbox}
\title{Plot a box (with x).}
\usage{
xbox(col = Seeblau, dim = c(1, 1))
}
\arguments{
\item{col}{The color to fill the box with (i.e., its background color).
Default: \code{col = unlist(seeblau)}.}

\item{dim}{The x- and y-dimensions of the box. 
Default: \code{dim = c(1, 1)} (i.e., a unit square).}
}
\description{
\code{xbox} plots a box with a cross (x) in its top-right corner.
}
\details{
The cross (x) appears rectangular when viewing the plot 
at the correct aspect ratio (as defined by \code{dim}).
}
\examples{
xbox()  # default box

# Options:
xbox(col = unlist(bordeaux))
xbox(dim = c(2, 1)) # 2:1 dimension (wider than high)

}
\seealso{
\code{\link{slide}} to plot a slide (or frame).

Other plot functions: \code{\link{slide}}
}
\concept{plot functions}
