% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box_calls.R
\name{xbox}
\alias{xbox}
\title{Plot a box (with x)}
\usage{
xbox(col = Seeblau, dim = c(1, 1), use_x = TRUE)
}
\arguments{
\item{col}{The color to fill the box with (i.e., its background color).
Default: \code{col = Seeblau}.}

\item{dim}{The x- and y-dimensions of the box (as numeric). 
Default: \code{dim = c(1, 1)} (i.e., a unit square).}

\item{use_x}{Plot a cross in upper right corner (as logical)? 
Default: \code{use_x = TRUE}.}
}
\description{
\code{xbox} plots a box with a cross (x) in its top-right corner.
}
\details{
The cross (x) appears rectangular when viewing the plot 
at the correct aspect ratio (as defined by \code{dim}).
}
\examples{
xbox()  # default box

# Options:
xbox(col = Bordeaux)
xbox(dim = c(2, 1)) # 2:1 dimension (wider than high)

}
\seealso{
\code{\link{post}} to add text to an xbox;   
\code{\link{slide}} to plot a new slide (or frame).

Other plot functions: 
\code{\link{slide}()},
\code{\link{theme_bwkn}()},
\code{\link{theme_grau}()},
\code{\link{theme_minikn}()},
\code{\link{theme_unikn}()}
}
\concept{plot functions}
